bit more structure as we go
