+++ server/mongoose.js
@@ -1,0 +1,5 @@
+const mongoose = require('mongoose');
+
+mongoose.connect('mongodb://localhost/bookclub');
+
+module.exports = mongoose;
