+++ server/samples.js
@@ -1,0 +1,6801 @@
+const Books = require('./books.js');
+
+Books.count()
+  .limit(1)
+  .then(hasBooks => {
+    if (!hasBooks) {
+      Books.insert(sampleBooks);
+    }
+  });
+
+const sampleBooks = [
+  {
+    title_suggest: 'Dansu, dansu, dansu',
+    edition_key: [
+      'OL14789568M',
+      'OL13567561M',
+      'OL10201956M',
+      'OL22271127M',
+      'OL1405842M',
+      'OL26231082M',
+      'OL1109338M',
+      'OL17383101M',
+      'OL23045939M',
+      'OL23023967M',
+      'OL8626357M',
+      'OL9272836M',
+      'OL24367381M',
+      'OL7318181M',
+      'OL24367383M',
+      'OL12545873M',
+      'OL9347690M',
+    ],
+    cover_i: 7886358,
+    id_librarything: ['5006', '8514443'],
+    has_fulltext: true,
+    text: [
+      'OL14789568M',
+      'OL13567561M',
+      'OL10201956M',
+      'OL22271127M',
+      'OL1405842M',
+      'OL26231082M',
+      'OL1109338M',
+      'OL17383101M',
+      'OL23045939M',
+      'OL23023967M',
+      'OL8626357M',
+      'OL9272836M',
+      'OL24367381M',
+      'OL7318181M',
+      'OL24367383M',
+      'OL12545873M',
+      'OL9347690M',
+      '5367002145',
+      '9621710960',
+      '0679753796',
+      '4770016832',
+      '9785367002140',
+      '4061850067',
+      '0099448769',
+      '0140154361',
+      '9789621710956',
+      '9626344350',
+      '9784062041225',
+      '9780140154368',
+      '1860469531',
+      '3832155333',
+      '4062041227',
+      '9784061850057',
+      '9783442730742',
+      '9782020664080',
+      '9784770016836',
+      '9780679753797',
+      '9780099448761',
+      '9785942781217',
+      '4061850059',
+      '9789621710963',
+      '5942781214',
+      '0241131456',
+      '9781860469534',
+      '9780241131459',
+      '3442730740',
+      '2020664089',
+      '9621710952',
+      '9789626344354',
+      '9783832155339',
+      '5942782261',
+      '9785942782269',
+      '9784061850064',
+      'Murakami Haruki',
+      'Kovalenin, Dmitri\u00ed.',
+      'Birnbaum, Alfred.',
+      'Rupert Degas (Narrator)',
+      '220562471',
+      '32709465',
+      '76472411',
+      '28218488',
+      '223430469',
+      'dansudansudansu00mura',
+      'dancedancedancen00mura',
+      'OL382524A',
+      'In library',
+      'Accessible book',
+      'Magical Realism',
+      'Fiction',
+      'Protected DAISY',
+      'Domestic fiction',
+      'roman',
+      'a novel',
+      'Dansu, dansu, dansu',
+      '/works/OL2625539W',
+      'Haruki Murakami ; [perevod s i\ufe20a\ufe21ponskogo D.V. Kovalenina].',
+      'Cunshang Chunshu zhu ; Ye Hui yi.',
+      'by Haruki Murakami ; translated by Alfred Birnbaum.',
+      'Murakami Haruki.',
+      'Bo yi chu ban',
+      'Vintage',
+      'Hamish Hamilton',
+      'btb',
+      'Vintage Books',
+      'Naxos AudioBooks',
+      'Kodansha International',
+      'Kodansha, Advanced Readers Copy',
+      'K\u014ddansha',
+      'Editions du Seuil',
+      'DuMont',
+      'The Harvill Press',
+      'Distributed in the U.S. by Kodansha America',
+      'Amfora',
+      'Penguin',
+      'Danse, danse, danse',
+      'Dansu dansu dansu.',
+      'Dance, Dance, Dance',
+      'Wu, wu, wu ba',
+      'D\u0117ns, d\u0117ns, d\u0117ns',
+      'Dansu, dansu, dansu.',
+      'Dance, dance, dance--',
+      'Dance dance dance',
+      'Tanz mit dem Schafsmann',
+      'Dance Dance Dance',
+      'Dance, dance, dance',
+      '93014098',
+      '94034713',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'I often dream about the Dolphin Hotel.',
+      'Japan',
+      '1980s',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: [
+      'Kovalenin, Dmitri\u00ed.',
+      'Birnbaum, Alfred.',
+      'Rupert Degas (Narrator)',
+    ],
+    ia_loaded_id: ['dansudansudansu00mura'],
+    seed: [
+      '/books/OL14789568M',
+      '/books/OL13567561M',
+      '/books/OL10201956M',
+      '/books/OL22271127M',
+      '/books/OL1405842M',
+      '/books/OL26231082M',
+      '/books/OL1109338M',
+      '/books/OL17383101M',
+      '/books/OL23045939M',
+      '/books/OL23023967M',
+      '/books/OL8626357M',
+      '/books/OL9272836M',
+      '/books/OL24367381M',
+      '/books/OL7318181M',
+      '/books/OL24367383M',
+      '/books/OL12545873M',
+      '/books/OL9347690M',
+      '/works/OL2625539W',
+      '/subjects/fiction',
+      '/subjects/protected_daisy',
+      '/subjects/domestic_fiction',
+      '/subjects/in_library',
+      '/subjects/magical_realism',
+      '/subjects/accessible_book',
+      '/subjects/place:japan',
+      '/subjects/time:1980s',
+      '/authors/OL382524A',
+    ],
+    oclc: ['220562471', '32709465', '76472411', '28218488', '223430469'],
+    ia: ['dansudansudansu00mura', 'dancedancedancen00mura'],
+    isbn: [
+      '5367002145',
+      '9621710960',
+      '0679753796',
+      '4770016832',
+      '9785367002140',
+      '4061850067',
+      '0099448769',
+      '0140154361',
+      '9789621710956',
+      '9626344350',
+      '9784062041225',
+      '9780140154368',
+      '1860469531',
+      '3832155333',
+      '4062041227',
+      '9784061850057',
+      '9783442730742',
+      '9782020664080',
+      '9784770016836',
+      '9780679753797',
+      '9780099448761',
+      '9785942781217',
+      '4061850059',
+      '9789621710963',
+      '5942781214',
+      '0241131456',
+      '9781860469534',
+      '9780241131459',
+      '3442730740',
+      '2020664089',
+      '9621710952',
+      '9789626344354',
+      '9783832155339',
+      '5942782261',
+      '9785942782269',
+      '9784061850064',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'In library',
+      'Accessible book',
+      'Magical Realism',
+      'Fiction',
+      'Protected DAISY',
+      'Domestic fiction',
+    ],
+    title: 'Dansu, dansu, dansu',
+    lending_identifier_s: 'dansudansudansu00mura',
+    ia_collection_s:
+      'printdisabled;internetarchivebooks;inlibrary;china;browserlending',
+    publish_date: [
+      '1991',
+      '1995',
+      '1994',
+      'February 7, 2002',
+      '1988',
+      '2002',
+      '2003',
+      'November 1, 2007',
+      '2001',
+      '2006',
+    ],
+    type: 'work',
+    ebook_count_i: 2,
+    publish_place: [
+      'Tokyo',
+      'Xianggang',
+      'France',
+      'New York',
+      'London',
+      'Cologne, Germany',
+      'Sankt-Peterburg',
+      'Munich, Germany',
+      'T\u014dky\u014d',
+    ],
+    ia_box_id: ['IA175001', 'IA102801'],
+    edition_count: 17,
+    first_publish_year: 1988,
+    key: '/works/OL2625539W',
+    id_goodreads: [
+      '903153',
+      '978096',
+      '266571',
+      '2376535',
+      '2376566',
+      '978094',
+      '17800',
+      '334598',
+      '45309',
+      '2376552',
+      '978095',
+    ],
+    public_scan_b: false,
+    publisher: [
+      'Bo yi chu ban',
+      'Vintage',
+      'Hamish Hamilton',
+      'btb',
+      'Vintage Books',
+      'Naxos AudioBooks',
+      'Kodansha International',
+      'Kodansha, Advanced Readers Copy',
+      'K\u014ddansha',
+      'Editions du Seuil',
+      'DuMont',
+      'The Harvill Press',
+      'Distributed in the U.S. by Kodansha America',
+      'Amfora',
+      'Penguin',
+    ],
+    language: ['ger', 'fre', 'eng', 'chi', 'jpn', 'rus'],
+    lccn: ['93014098', '94034713'],
+    last_modified_i: 1488132195,
+    lending_edition_s: 'OL13567561M',
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL1405842M',
+    first_sentence: ['I often dream about the Dolphin Hotel.'],
+    publish_year: [1991, 1995, 1994, 1988, 2002, 2003, 2001, 2006, 2007],
+    printdisabled_s: 'OL1405842M;OL13567561M',
+    place: ['Japan'],
+    time: ['1980s'],
+    id_dnb: ['964220768', '965951030'],
+  },
+  {
+    title_suggest: 'Kafka sur le Rivage',
+    edition_key: [
+      'OL19617410M',
+      'OL23086017M',
+      'OL23059586M',
+      'OL8931367M',
+      'OL8931327M',
+      'OL24277007M',
+      'OL9453775M',
+      'OL8363016M',
+      'OL9161080M',
+      'OL8364133M',
+      'OL26420371M',
+      'OL24367154M',
+    ],
+    cover_i: 746985,
+    id_librarything: ['5819'],
+    has_fulltext: true,
+    text: [
+      'OL19617410M',
+      'OL23086017M',
+      'OL23059586M',
+      'OL8931367M',
+      'OL8931327M',
+      'OL24277007M',
+      'OL9453775M',
+      'OL8363016M',
+      'OL9161080M',
+      'OL8364133M',
+      'OL26420371M',
+      'OL24367154M',
+      '9785551416395',
+      '9789626344057',
+      '9781400079278',
+      '1400043662',
+      '9781400043668',
+      '9785699109388',
+      '1843432684',
+      '9783442740437',
+      '9781843432685',
+      '4103534133',
+      '5699109382',
+      '9784103534143',
+      '9626344059',
+      '9781400044818',
+      '3442740436',
+      '5551416399',
+      '2264056169',
+      '1400079276',
+      '9784103534136',
+      '4103534141',
+      '1400044812',
+      '1843431106',
+      '9781843431107',
+      '9782264056160',
+      'Murakami Haruki',
+      'Oliver Le Sueur (Narrator)',
+      'Sean Barrett (Narrator)',
+      '232242470',
+      '57559842',
+      '50764038',
+      '458686407',
+      'kafkaonshore00mura_0',
+      'OL382524A',
+      'Runaway teenagers',
+      'Literature',
+      'Translations into Russian',
+      'Japanese fiction',
+      'Accessible book',
+      'Fiction',
+      'Parent and child',
+      'OverDrive',
+      'Protected DAISY',
+      'Internet Archive Wishlist',
+      'Kafka on the shore.',
+      'roman',
+      'Kafka on the shore',
+      'Kafka sur le Rivage',
+      '/works/OL2625537W',
+      'Kharuki Murakami ; perevod s i\ufe20a\ufe21ponskogo Ivana i Sergei\ufe20a\ufe21  Logachevykh.',
+      'Murakami Haruki.',
+      'Knopf',
+      'Shinch\u014dsha',
+      'Vintage',
+      'Naxos of America',
+      '10/18',
+      '\u0116KSMO',
+      'btb',
+      'The Harvill Press',
+      'VINTAGE (RAND)',
+      'Knopf Doubleday Publishing Group',
+      'Kafka on the shore',
+      'Kafka on the Shore',
+      'Umibe-no Kafuka.',
+      'Kafka am Strand',
+      'Umibe no Kafuka =',
+      'Kafka na pli\ufe20a\ufe21zhe',
+      'Umibe no Kafuka',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Oliver Le Sueur (Narrator)', 'Sean Barrett (Narrator)'],
+    ia_loaded_id: ['kafkaonshore00mura_0'],
+    seed: [
+      '/books/OL19617410M',
+      '/books/OL23086017M',
+      '/books/OL23059586M',
+      '/books/OL8931367M',
+      '/books/OL8931327M',
+      '/books/OL24277007M',
+      '/books/OL9453775M',
+      '/books/OL8363016M',
+      '/books/OL9161080M',
+      '/books/OL8364133M',
+      '/books/OL26420371M',
+      '/books/OL24367154M',
+      '/works/OL2625537W',
+      '/subjects/runaway_teenagers',
+      '/subjects/literature',
+      '/subjects/translations_into_russian',
+      '/subjects/japanese_fiction',
+      '/subjects/fiction',
+      '/subjects/parent_and_child',
+      '/subjects/overdrive',
+      '/subjects/protected_daisy',
+      '/subjects/internet_archive_wishlist',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    oclc: ['232242470', '57559842', '50764038', '458686407'],
+    ia: ['kafkaonshore00mura_0'],
+    isbn: [
+      '9785551416395',
+      '9789626344057',
+      '9781400079278',
+      '1400043662',
+      '9781400043668',
+      '9785699109388',
+      '1843432684',
+      '9783442740437',
+      '9781843432685',
+      '4103534133',
+      '5699109382',
+      '9784103534143',
+      '9626344059',
+      '9781400044818',
+      '3442740436',
+      '5551416399',
+      '2264056169',
+      '1400079276',
+      '9784103534136',
+      '4103534141',
+      '1400044812',
+      '1843431106',
+      '9781843431107',
+      '9782264056160',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Runaway teenagers',
+      'Literature',
+      'Translations into Russian',
+      'Japanese fiction',
+      'Accessible book',
+      'Fiction',
+      'Parent and child',
+      'OverDrive',
+      'Protected DAISY',
+      'Internet Archive Wishlist',
+    ],
+    title: 'Kafka sur le Rivage',
+    ia_collection_s: 'printdisabled;china;internetarchivebooks',
+    publish_date: [
+      'January 18, 2005',
+      '2002',
+      'January 3, 2006',
+      'May 2006',
+      '2005',
+      'January 6, 2005',
+      '2009',
+      '2007',
+    ],
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: [
+      'New York',
+      'London, United Kingdom',
+      'Munich, Germany',
+      'T\u014dky\u014d',
+      'Paris, France',
+      'Moskva',
+    ],
+    ia_box_id: ['IA129801'],
+    edition_count: 12,
+    key: '/works/OL2625537W',
+    id_goodreads: [
+      '463050',
+      '959123',
+      '959124',
+      '1936621',
+      '1936622',
+      '4929',
+      '959126',
+      '3646541',
+      '45314',
+    ],
+    public_scan_b: false,
+    id_overdrive: ['A8B50A2F-F2C0-4CF7-9F4F-80BB2272042A'],
+    publisher: [
+      'Knopf',
+      'Shinch\u014dsha',
+      'Vintage',
+      'Naxos of America',
+      '10/18',
+      '\u0116KSMO',
+      'btb',
+      'The Harvill Press',
+      'VINTAGE (RAND)',
+      'Knopf Doubleday Publishing Group',
+    ],
+    language: ['rus', 'ger', 'eng', 'fre', 'jpn'],
+    last_modified_i: 1516912296,
+    first_publish_year: 2002,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL8364133M',
+    publish_year: [2002, 2009, 2006, 2007, 2005],
+    printdisabled_s: 'OL8363016M',
+    place: ['Japan'],
+    id_dnb: ['993844944'],
+  },
+  {
+    title_suggest: 'The wind-up bird chronicle',
+    edition_key: [
+      'OL12355321M',
+      'OL16359913M',
+      'OL657653M',
+      'OL18192187M',
+      'OL21503461M',
+      'OL7701131M',
+      'OL8626188M',
+      'OL23064119M',
+      'OL15515237M',
+      'OL23120715M',
+      'OL9161092M',
+    ],
+    cover_i: 6253176,
+    id_librarything: ['5069', '8514416'],
+    has_fulltext: true,
+    text: [
+      'OL12355321M',
+      'OL16359913M',
+      'OL657653M',
+      'OL18192187M',
+      'OL21503461M',
+      'OL7701131M',
+      'OL8626188M',
+      'OL23064119M',
+      'OL15515237M',
+      'OL23120715M',
+      'OL9161092M',
+      '0679775439',
+      '4101001413',
+      '9781860464706',
+      '186046470X',
+      '9789626344187',
+      '0099448793',
+      '5867121356',
+      '4103534036',
+      '0679446699',
+      '9784101001418',
+      '410100143X',
+      '9780679446699',
+      '4101001421',
+      '9626344180',
+      '9784101001432',
+      '1860465811',
+      '9781860465819',
+      '9784101001425',
+      '9784103534037',
+      '9785867121358',
+      '9780099448792',
+      '9780679775430',
+      'Murakami Haruki',
+      'Logachev, Ivan.',
+      'Trsan, Tinen Cao \u0110\u0103ng.',
+      'Rupert Degas (Narrator)',
+      'Logachev, Serge\u01d0.',
+      'Rubin, Jay, 1941-',
+      '39915729',
+      '60155085',
+      '50543227',
+      'windupbirdchroni00mura',
+      'biennienkychimvq00mura',
+      'OL382524A',
+      'Politics and government',
+      'In library',
+      'Man-woman relationships',
+      'Japanese fiction',
+      'Accessible book',
+      'Translations into English',
+      'Fiction',
+      'Marriage',
+      'Protected DAISY',
+      'Missing persons',
+      'translated from the Japanese ; by Jay Rubin.',
+      'A Novel',
+      'roman',
+      'The wind-up bird chronicle',
+      '/works/OL2625494W',
+      'Kharuki Murakami ; [per. s i\ufe20a\ufe21ponskogo  Ivana i Sergei\ufe20a\ufe21 Logachevykh].',
+      'Murakami, Haruki ; Trsan Tinen Cao \u0110\u0103ng d\u1ecbch.',
+      'Murakami Haruki.',
+      '[by] Haruki Murakami ; translated from the Japanese by Jay Rubin.',
+      'Haruki Murakami :',
+      'Haruki Murakami ; translated from the Japanese by Jay Rubin.',
+      'Alfred A. Knopf',
+      'Vintage International',
+      'C\u00f4ng ty v\u0103n h\u00f3a & truysen th\u00f4ng Nh\u00e3 Nam',
+      'Vintage',
+      'NXB Huoi nh\u00e0 v\u0103n',
+      'Shinchosha',
+      'Harvill Press',
+      'Naxos Audiobooks',
+      'Nezavisimai\ufe20a\ufe21 gazeta',
+      'Shinch\u014dsha',
+      'The Harvill Press',
+      'Wind-up bird chronicle',
+      'Nejimaki-dori kuronikuru.',
+      'Nejimaki-dori kuronikuru',
+      'The Wind-Up Bird Chronicle',
+      'Khroniki zavodno\u01d0 ptit\ufe20s\ufe21y',
+      'Wind-Up Bird Chronicle',
+      'The Wind-up Bird Chronicle (The Complete Classics)',
+      'Bi\u00ean ni\u00ean k\u00fd chim vqan d\u00e2y c\u00f3t',
+      '97002813',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      "When the phone rang I was in the kitchen, boiling a potful of spaghetti and whistling along with an FM broadcast of the overture of the Rossini's The Thieving Magpie, which has to be the perfect music for cooking pasta.",
+      'Japan',
+      '20th century',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: [
+      'Logachev, Ivan.',
+      'Trsan, Tinen Cao \u0110\u0103ng.',
+      'Rupert Degas (Narrator)',
+      'Logachev, Serge\u01d0.',
+      'Rubin, Jay, 1941-',
+    ],
+    ia_loaded_id: ['biennienkychimvq00mura'],
+    seed: [
+      '/books/OL12355321M',
+      '/books/OL16359913M',
+      '/books/OL657653M',
+      '/books/OL18192187M',
+      '/books/OL21503461M',
+      '/books/OL7701131M',
+      '/books/OL8626188M',
+      '/books/OL23064119M',
+      '/books/OL15515237M',
+      '/books/OL23120715M',
+      '/books/OL9161092M',
+      '/works/OL2625494W',
+      '/subjects/accessible_book',
+      '/subjects/fiction',
+      '/subjects/japanese_fiction',
+      '/subjects/man-woman_relationships',
+      '/subjects/marriage',
+      '/subjects/missing_persons',
+      '/subjects/politics_and_government',
+      '/subjects/protected_daisy',
+      '/subjects/translations_into_english',
+      '/subjects/in_library',
+      '/subjects/place:japan',
+      '/subjects/time:20th_century',
+      '/authors/OL382524A',
+    ],
+    oclc: ['39915729', '60155085', '50543227'],
+    ia: ['windupbirdchroni00mura', 'biennienkychimvq00mura'],
+    isbn: [
+      '0679775439',
+      '4101001413',
+      '9781860464706',
+      '186046470X',
+      '9789626344187',
+      '0099448793',
+      '5867121356',
+      '4103534036',
+      '0679446699',
+      '9784101001418',
+      '410100143X',
+      '9780679446699',
+      '4101001421',
+      '9626344180',
+      '9784101001432',
+      '1860465811',
+      '9781860465819',
+      '9784101001425',
+      '9784103534037',
+      '9785867121358',
+      '9780099448792',
+      '9780679775430',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Politics and government',
+      'In library',
+      'Man-woman relationships',
+      'Japanese fiction',
+      'Accessible book',
+      'Translations into English',
+      'Fiction',
+      'Marriage',
+      'Protected DAISY',
+      'Missing persons',
+    ],
+    title: 'The wind-up bird chronicle',
+    lending_identifier_s: 'biennienkychimvq00mura',
+    ia_collection_s:
+      'printdisabled;americana;internetarchivebooks;inlibrary;china;browserlending',
+    publish_date: [
+      '1994',
+      '1997',
+      'September 1, 1998',
+      '1998',
+      'May 22, 2007',
+      'May 1999',
+      '2002',
+      '2003',
+      '2006',
+    ],
+    type: 'work',
+    ebook_count_i: 2,
+    publish_place: [
+      'H\u00e0 Nuoi',
+      'T\u014dky\u014d',
+      'London',
+      'New York',
+      'Moskva',
+    ],
+    ia_box_id: ['IA105811', 'IA182701', 'IA155901'],
+    edition_count: 11,
+    first_publish_year: 1994,
+    key: '/works/OL2625494W',
+    id_goodreads: [
+      '3742284',
+      '821684',
+      '70933',
+      '151512',
+      '489283',
+      '284066',
+      '45312',
+      '1633688',
+      '1633692',
+      '1633690',
+      '11273',
+    ],
+    public_scan_b: false,
+    publisher: [
+      'Alfred A. Knopf',
+      'Vintage International',
+      'C\u00f4ng ty v\u0103n h\u00f3a & truysen th\u00f4ng Nh\u00e3 Nam',
+      'Vintage',
+      'NXB Huoi nh\u00e0 v\u0103n',
+      'Shinchosha',
+      'Harvill Press',
+      'Naxos Audiobooks',
+      'Nezavisimai\ufe20a\ufe21 gazeta',
+      'Shinch\u014dsha',
+      'The Harvill Press',
+    ],
+    language: ['vie', 'rus', 'eng', 'jpn'],
+    lccn: ['97002813'],
+    last_modified_i: 1497169112,
+    lending_edition_s: 'OL23120715M',
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL21503461M',
+    first_sentence: [
+      "When the phone rang I was in the kitchen, boiling a potful of spaghetti and whistling along with an FM broadcast of the overture of the Rossini's The Thieving Magpie, which has to be the perfect music for cooking pasta.",
+    ],
+    publish_year: [1994, 1997, 1999, 1998, 2002, 2003, 2006, 2007],
+    printdisabled_s: 'OL23120715M;OL21503461M',
+    place: ['Japan'],
+    time: ['20th century'],
+  },
+  {
+    title_suggest: 'Hitsuji o meguru b\u014dken',
+    edition_key: [
+      'OL3235422M',
+      'OL2068263M',
+      'OL1879547M',
+      'OL23750122M',
+      'OL14971455M',
+      'OL23041656M',
+      'OL23003169M',
+      'OL7318182M',
+      'OL9272837M',
+      'OL7426704M',
+      'OL9161088M',
+    ],
+    cover_i: 7891404,
+    id_librarything: ['4987'],
+    has_fulltext: true,
+    text: [
+      'OL3235422M',
+      'OL2068263M',
+      'OL1879547M',
+      'OL23750122M',
+      'OL14971455M',
+      'OL23041656M',
+      'OL23003169M',
+      'OL7318182M',
+      'OL9272837M',
+      'OL7426704M',
+      'OL9161088M',
+      '9780870119057',
+      '4062002418',
+      '9780375718946',
+      '0870119052',
+      '9780241129944',
+      '9780099448778',
+      '0452265169',
+      '9621710731',
+      '9789626344149',
+      '0099448777',
+      '037571894X',
+      '9789621710727',
+      '576840502X',
+      '9785768405021',
+      '9789621710734',
+      '9784062002417',
+      '9626344148',
+      '9621710723',
+      '024112994X',
+      '9780452265165',
+      'Murakami Haruki',
+      'Rupert Degas (Narrator)',
+      '19670739',
+      '44002560',
+      'wildsheepchase00murarich',
+      'wildsheepchase00haru',
+      'wildsheepchase00mura',
+      'OL382524A',
+      'Japanese fiction',
+      'Accessible book',
+      'Translations into Russian',
+      'In library',
+      'Protected DAISY',
+      'A Novel',
+      'roman',
+      'a novel',
+      'Hitsuji o meguru b\u014dken',
+      '/works/OL2625493W',
+      'Haruki Murakami ; perevod s i\ufe20a\ufe21ponskogo Dm. Kovalenina.',
+      'Haruki Murakami ; translated by Alfred Birnbaum.',
+      'by Haruki Murakami ; translated by Alfred Birnbaum.',
+      'Cunshang Chunshu ; Ye Hui yi.',
+      'Murakami Haruki.',
+      'Vintage',
+      'Kodansha International',
+      'Naxos Audiobooks',
+      '"Azbuka"',
+      'Penguin Books',
+      'Hamish Hamilton',
+      'Bo yi chu ban ji tuan',
+      'K\u014ddansha',
+      'Wild sheep chase',
+      'A wild sheep chase',
+      'A Wild Sheep Chase (Junior Classics)',
+      'Hitsuji o meguru b\u014dken.',
+      'Xun yang de mao xian',
+      'Okhota na ovet\ufe20s\ufe21',
+      'Hitsuji o meguru b\u014dken',
+      'Hitsuji o meguru b\u014dken.',
+      'A Wild Sheep Chase',
+      '88080299',
+      '90039976',
+      '83149162',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'It was a short one-paragraph item in the morning edition.',
+      '20th century',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Rupert Degas (Narrator)'],
+    seed: [
+      '/books/OL3235422M',
+      '/books/OL2068263M',
+      '/books/OL1879547M',
+      '/books/OL23750122M',
+      '/books/OL14971455M',
+      '/books/OL23041656M',
+      '/books/OL23003169M',
+      '/books/OL7318182M',
+      '/books/OL9272837M',
+      '/books/OL7426704M',
+      '/books/OL9161088M',
+      '/works/OL2625493W',
+      '/subjects/accessible_book',
+      '/subjects/japanese_fiction',
+      '/subjects/protected_daisy',
+      '/subjects/translations_into_russian',
+      '/subjects/in_library',
+      '/subjects/time:20th_century',
+      '/authors/OL382524A',
+    ],
+    oclc: ['19670739', '44002560'],
+    ia: [
+      'wildsheepchase00murarich',
+      'wildsheepchase00haru',
+      'wildsheepchase00mura',
+    ],
+    isbn: [
+      '9780870119057',
+      '4062002418',
+      '9780375718946',
+      '0870119052',
+      '9780241129944',
+      '9780099448778',
+      '0452265169',
+      '9621710731',
+      '9789626344149',
+      '0099448777',
+      '037571894X',
+      '9789621710727',
+      '576840502X',
+      '9785768405021',
+      '9789621710734',
+      '9784062002417',
+      '9626344148',
+      '9621710723',
+      '024112994X',
+      '9780452265165',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Japanese fiction',
+      'Accessible book',
+      'Translations into Russian',
+      'In library',
+      'Protected DAISY',
+    ],
+    title: 'Hitsuji o meguru b\u014dken',
+    lending_identifier_s: 'wildsheepchase00murarich',
+    ia_collection_s:
+      'printdisabled;americana;internetarchivebooks;inlibrary;china;browserlending',
+    printdisabled_s: 'OL23750122M;OL7426704M;OL2068263M',
+    type: 'work',
+    ebook_count_i: 3,
+    publish_place: [
+      'New York, N.Y., U.S.A',
+      'Tokyo',
+      'Xianggang',
+      'London',
+      'Toyyo, New York',
+      'Sankt-Peterburg',
+      'New York, N.Y',
+      'T\u014dky\u014d',
+    ],
+    ia_box_id: ['IA101320', 'IA157801', 'IA138109'],
+    edition_count: 11,
+    first_publish_year: 1982,
+    key: '/works/OL2625493W',
+    id_goodreads: [
+      '912372',
+      '485875',
+      '818111',
+      '146927',
+      '11298',
+      '3107333',
+      '3107334',
+    ],
+    public_scan_b: false,
+    publisher: [
+      'Vintage',
+      'Kodansha International',
+      'Naxos Audiobooks',
+      '"Azbuka"',
+      'Penguin Books',
+      'Hamish Hamilton',
+      'Bo yi chu ban ji tuan',
+      'K\u014ddansha',
+    ],
+    language: ['chi', 'rus', 'eng', 'jpn'],
+    lccn: ['88080299', '90039976', '83149162'],
+    last_modified_i: 1486042678,
+    lending_edition_s: 'OL2068263M',
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL2068263M',
+    first_sentence: [
+      'It was a short one-paragraph item in the morning edition.',
+    ],
+    publish_year: [1990, 1992, 1982, 1998, 1989, 2002, 2000, 2006],
+    publish_date: [
+      '1990',
+      '1992',
+      '1982',
+      'April 20, 2000',
+      '1998',
+      '1989',
+      'April 9, 2002',
+      'July 30, 2006',
+    ],
+    time: ['20th century'],
+  },
+  {
+    title_suggest: 'Sekai no owari to h\u0101do-boirudo wand\u0101rando',
+    edition_key: [
+      'OL22459192M',
+      'OL23087755M',
+      'OL1766393M',
+      'OL1544712M',
+      'OL14949205M',
+      'OL1745775M',
+      'OL8626341M',
+      'OL7318183M',
+      'OL9272838M',
+      'OL23015214M',
+    ],
+    subtitle: 'a novel',
+    has_fulltext: true,
+    id_librarything: ['5798', '8514461', '9229178'],
+    text: [
+      'OL22459192M',
+      'OL23087755M',
+      'OL1766393M',
+      'OL1544712M',
+      'OL14949205M',
+      'OL1745775M',
+      'OL8626341M',
+      'OL7318183M',
+      'OL9272838M',
+      'OL23015214M',
+      'a novel',
+      'Murakami Haruki',
+      'Kovalenin, Dmitri\u012d.',
+      '55505601',
+      '26399689',
+      '60009350',
+      '15331273',
+      'hardboiledwonder00mura',
+      '4101001340',
+      '4770015445',
+      '9784770015440',
+      '569902784X',
+      '4106006448',
+      '4101001359',
+      '9781860469053',
+      '9785699027842',
+      '0679743464',
+      '9784101001357',
+      '1860469051',
+      '9780679743460',
+      '0099448785',
+      '0241131448',
+      '9780099448785',
+      '9784101001340',
+      '9780140154351',
+      '9784106006449',
+      '0140154353',
+      '9780241131442',
+      'OL382524A',
+      'Fantasy fiction',
+      'Accessible book',
+      'Protected DAISY',
+      'Sekai no owari to h\u0101do-boirudo wand\u0101rando',
+      '/works/OL15338506W',
+      'Haruki Murakami ; translated by Alfred Birnbaum.',
+      'Murakami Haruki cho.',
+      'Haruki Murakami ; [per. s i\ufe20a\ufe21ponskogo D. Kovalenina].',
+      'Murakami Haruki.',
+      'Shinch\u014dsha',
+      'Vintage',
+      'Hamish Hamilton',
+      'Vintage Books',
+      '"\u0116KSMO"',
+      'Penguin Books USA',
+      'Kodansha International',
+      'The Harvill Press',
+      'Distributed in the U.S. by Kodansha America',
+      'Penguin',
+      'Hardboiled Wonderland and the End of the World',
+      'Sekai no owari to h\u0101do-boirudo wand\u0101rando',
+      'The hard-boiled wonderland and the end of the world',
+      'Sekai no owari to hard-boliled wonderland.',
+      'Sekai no owari to h\u0101do-boirudo wand\u0101rando.',
+      'Strana chudes bez tormozov i Konet\ufe20s\ufe21 Sveta',
+      'Hard-boiled Wonderland and the End of the World',
+      'Hard-boiled wonderland and the end of the world',
+      '92056345',
+      '91023818',
+      '92116678',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'THE elevator continued its impossible slow ascent.',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Kovalenin, Dmitri\u012d.'],
+    seed: [
+      '/books/OL22459192M',
+      '/books/OL23087755M',
+      '/books/OL1766393M',
+      '/books/OL1544712M',
+      '/books/OL14949205M',
+      '/books/OL1745775M',
+      '/books/OL8626341M',
+      '/books/OL7318183M',
+      '/books/OL9272838M',
+      '/books/OL23015214M',
+      '/works/OL15338506W',
+      '/subjects/fantasy_fiction',
+      '/authors/OL382524A',
+    ],
+    oclc: ['55505601', '26399689', '60009350', '15331273'],
+    ia: ['hardboiledwonder00mura'],
+    isbn: [
+      '4101001340',
+      '4770015445',
+      '9784770015440',
+      '569902784X',
+      '4106006448',
+      '4101001359',
+      '9781860469053',
+      '9785699027842',
+      '0679743464',
+      '9784101001357',
+      '1860469051',
+      '9780679743460',
+      '0099448785',
+      '0241131448',
+      '9780099448785',
+      '9784101001340',
+      '9780140154351',
+      '9784106006449',
+      '0140154353',
+      '9780241131442',
+    ],
+    author_key: ['OL382524A'],
+    subject: ['Fantasy fiction', 'Accessible book', 'Protected DAISY'],
+    title: 'Sekai no owari to h\u0101do-boirudo wand\u0101rando',
+    ia_collection_s: 'printdisabled;china;internetarchivebooks',
+    printdisabled_s: 'OL1745775M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: [
+      'Tokyo',
+      'New York',
+      'London',
+      'New York, N.Y., USA',
+      'T\u014dky\u014d',
+      'Moskva',
+    ],
+    ia_box_id: ['IA198401'],
+    edition_count: 10,
+    key: '/works/OL15338506W',
+    id_goodreads: [
+      '833603',
+      '833605',
+      '833602',
+      '10374',
+      '1630199',
+      '1630198',
+      '1936596',
+      '106484',
+      '151511',
+      '45310',
+    ],
+    public_scan_b: false,
+    publisher: [
+      'Shinch\u014dsha',
+      'Vintage',
+      'Hamish Hamilton',
+      'Vintage Books',
+      '"\u0116KSMO"',
+      'Penguin Books USA',
+      'Kodansha International',
+      'The Harvill Press',
+      'Distributed in the U.S. by Kodansha America',
+      'Penguin',
+    ],
+    language: ['rus', 'eng', 'jpn'],
+    lccn: ['92056345', '91023818', '92116678'],
+    last_modified_i: 1496292283,
+    first_publish_year: 1985,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL1544712M',
+    first_sentence: ['THE elevator continued its impossible slow ascent.'],
+    publish_year: [1991, 1993, 1992, 1988, 2003, 2001, 1985],
+    publish_date: [
+      '1991',
+      '1993',
+      '1992',
+      'September 28, 2001',
+      '1988',
+      '2003',
+      '1985',
+      '2001',
+    ],
+  },
+  {
+    title_suggest: 'Tokio blues',
+    edition_key: ['OL9700086M', 'OL9144171M', 'OL25433373M'],
+    cover_i: 1048296,
+    isbn: [
+      '978-84-8383-504-3',
+      '8483103079',
+      '9788483103074',
+      '9871210191',
+      '8483835045',
+      '84-8310-307-9',
+      '9789871210190',
+    ],
+    has_fulltext: true,
+    id_depósito_legal: ['B. 3.110-2006'],
+    text: [
+      'OL9700086M',
+      'OL9144171M',
+      'OL25433373M',
+      '978-84-8383-504-3',
+      '8483103079',
+      '9788483103074',
+      '9871210191',
+      '8483835045',
+      '84-8310-307-9',
+      '9789871210190',
+      'Haruki Murakami',
+      'Lourdes Porta Fuentes (Translator)',
+      'tokioblues00mura',
+      'OL5036666A',
+      'Accessible book',
+      'Protected DAISY',
+      'Norwegian Wood',
+      'Tokio blues',
+      '/works/OL2625517W',
+      'TusQuets',
+      'Tusquets editores',
+      'Tokio Blues',
+    ],
+    author_name: ['Haruki Murakami'],
+    contributor: ['Lourdes Porta Fuentes (Translator)'],
+    seed: [
+      '/books/OL9700086M',
+      '/books/OL9144171M',
+      '/books/OL25433373M',
+      '/works/OL2625517W',
+      '/authors/OL5036666A',
+    ],
+    ia: ['tokioblues00mura'],
+    author_key: ['OL5036666A'],
+    subject: ['Accessible book', 'Protected DAISY'],
+    title: 'Tokio blues',
+    lending_identifier_s: 'tokioblues00mura',
+    ia_collection_s:
+      'printdisabled;browserlending;china;inlibrary;internetarchivebooks',
+    printdisabled_s: 'OL9144171M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['Barcelona, Spain'],
+    ia_box_id: ['IA1139403'],
+    edition_count: 3,
+    first_publish_year: 2005,
+    key: '/works/OL2625517W',
+    id_goodreads: ['291494', '11300'],
+    public_scan_b: false,
+    publisher: ['TusQuets', 'Tusquets editores'],
+    language: ['spa'],
+    last_modified_i: 1498289169,
+    lending_edition_s: 'OL9144171M',
+    id_librarything: ['5860'],
+    cover_edition_key: 'OL9144171M',
+    publish_year: [2006, 2005],
+    publish_date: ['December 30, 2005', 'September 2005', '2006'],
+  },
+  {
+    title_suggest: 'Colorless Tsukuru Tazaki and his years of pilgrimage',
+    publisher: ['Knopf', 'Vintage'],
+    cover_i: 7852099,
+    isbn: ['0099590379', '9780099590378', '0385352107', '9780385352109'],
+    has_fulltext: false,
+    title: 'Colorless Tsukuru Tazaki and his years of pilgrimage',
+    edition_key: ['OL26178712M', 'OL26318328M'],
+    last_modified_i: 1499984459,
+    edition_count: 2,
+    subject: ['Internet Archive Wishlist'],
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26178712M',
+    seed: [
+      '/books/OL26178712M',
+      '/books/OL26318328M',
+      '/works/OL17575566W',
+      '/subjects/internet_archive_wishlist',
+      '/authors/OL5036666A',
+    ],
+    first_publish_year: 2014,
+    publish_year: [2015, 2014],
+    key: '/works/OL17575566W',
+    text: [
+      'OL26178712M',
+      'OL26318328M',
+      '0099590379',
+      '9780099590378',
+      '0385352107',
+      '9780385352109',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Internet Archive Wishlist',
+      'Colorless Tsukuru Tazaki and his years of pilgrimage',
+      '/works/OL17575566W',
+      'Knopf',
+      'Vintage',
+      'Colorless Tsukuru Tazaki And His Years Of Pilgrimage A Novel',
+    ],
+    publish_date: ['2015', '2014'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'Los a\u00f1os de peregrinaci\u00f3n del chico sin color',
+    publisher: ['Tusquets', 'Tusquets '],
+    isbn: ['9789876701952', '8483837447', '9788483837443', '9876701959'],
+    has_fulltext: false,
+    edition_count: 2,
+    edition_key: ['OL26356764M', 'OL25435053M'],
+    last_modified_i: 1501173739,
+    title: 'Los a\u00f1os de peregrinaci\u00f3n del chico sin color',
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL25435053M',
+    seed: [
+      '/books/OL26356764M',
+      '/books/OL25435053M',
+      '/works/OL16810121W',
+      '/authors/OL5036666A',
+    ],
+    first_publish_year: 2013,
+    publish_year: [2013],
+    key: '/works/OL16810121W',
+    text: [
+      'OL26356764M',
+      'OL25435053M',
+      '9789876701952',
+      '8483837447',
+      '9788483837443',
+      '9876701959',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Los a\u00f1os de peregrinaci\u00f3n del chico sin color',
+      '/works/OL16810121W',
+      'Tusquets',
+      'Tusquets ',
+    ],
+    publish_date: ['2013'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'After the Quake',
+    edition_key: [
+      'OL8931302M',
+      'OL8626362M',
+      'OL7425273M',
+      'OL9272826M',
+      'OL7426647M',
+      'OL7318171M',
+      'OL9161105M',
+    ],
+    cover_i: 915106,
+    subtitle: 'Stories',
+    has_fulltext: false,
+    text: [
+      'OL8931302M',
+      'OL8626362M',
+      'OL7425273M',
+      'OL9272826M',
+      'OL7426647M',
+      'OL7318171M',
+      'OL9161105M',
+      'Stories',
+      'Murakami Haruki',
+      'Adam Sims (Narrator)',
+      'Rupert Degas (Narrator)',
+      'Teresa Gallagher (Narrator)',
+      'Jay Rubin (Translator)',
+      '9780375413902',
+      '9780099448563',
+      '0375713271',
+      '1860469671',
+      '1843430150',
+      '9781843430155',
+      '9789626344323',
+      '0375413901',
+      '9780375713279',
+      '9781860469671',
+      '9626344326',
+      '0099448564',
+      'OL382524A',
+      'Social life and customs',
+      'Japan in fiction',
+      'Fiction',
+      'Five straight days she spent in front of the television, staring at crumbled banks and hospitals, whole blocks of stores in flames, severed rail lines and expressways.',
+      'After the Quake',
+      'Japan',
+      '/works/OL2625533W',
+      'Knopf',
+      'Naxos of America',
+      'The Harvill Press',
+      'Harvill P.',
+      'Vintage',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      '20th century',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL8931302M',
+      '/books/OL8626362M',
+      '/books/OL7425273M',
+      '/books/OL9272826M',
+      '/books/OL7426647M',
+      '/books/OL7318171M',
+      '/books/OL9161105M',
+      '/works/OL2625533W',
+      '/subjects/social_life_and_customs',
+      '/subjects/japan_in_fiction',
+      '/subjects/fiction',
+      '/subjects/place:japan',
+      '/subjects/time:20th_century',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['31994'],
+    contributor: [
+      'Adam Sims (Narrator)',
+      'Rupert Degas (Narrator)',
+      'Teresa Gallagher (Narrator)',
+      'Jay Rubin (Translator)',
+    ],
+    isbn: [
+      '9780375413902',
+      '9780099448563',
+      '0375713271',
+      '1860469671',
+      '1843430150',
+      '9781843430155',
+      '9789626344323',
+      '0375413901',
+      '9780375713279',
+      '9781860469671',
+      '9626344326',
+      '0099448564',
+    ],
+    author_key: ['OL382524A'],
+    subject: ['Social life and customs', 'Japan in fiction', 'Fiction'],
+    title: 'After the Quake',
+    publish_date: [
+      'October 3, 2002',
+      'May 13, 2003',
+      '2002',
+      'August 13, 2002',
+      'March 6, 2003',
+      'May 22, 2007',
+    ],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 7,
+    key: '/works/OL2625533W',
+    id_goodreads: ['45308', '17809', '11299', '132762', '45313', '764310'],
+    publisher: [
+      'Knopf',
+      'Naxos of America',
+      'The Harvill Press',
+      'Harvill P.',
+      'Vintage',
+    ],
+    language: ['eng'],
+    last_modified_i: 1281583786,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL8626362M',
+    first_sentence: [
+      'Five straight days she spent in front of the television, staring at crumbled banks and hospitals, whole blocks of stores in flames, severed rail lines and expressways.',
+    ],
+    publish_year: [2002, 2003, 2007],
+    first_publish_year: 2002,
+    place: ['Japan'],
+    time: ['20th century'],
+  },
+  {
+    title_suggest: 'Hashiru koto ni tsuite kataru toki ni boku no kataru koto',
+    edition_key: [
+      'OL25620818M',
+      'OL23125654M',
+      'OL24270384M',
+      'OL16819451M',
+      'OL24241204M',
+      'OL23873280M',
+      'OL24095056M',
+    ],
+    cover_i: 7167416,
+    subtitle: 'a memoir',
+    has_fulltext: true,
+    id_librarything: ['2537238'],
+    text: [
+      'OL25620818M',
+      'OL23125654M',
+      'OL24270384M',
+      'OL16819451M',
+      'OL24241204M',
+      'OL23873280M',
+      'OL24095056M',
+      'a memoir',
+      'Murakami Haruki',
+      'Gabriel, Philip, 1953-',
+      'Gabriel, Philip.',
+      '246707220',
+      'whatitalkaboutwh00mura_812',
+      'whatitalkaboutwh00mura',
+      '978-9955-23-584-2',
+      '0307269191',
+      '9780307269478',
+      '0307389839',
+      '9780307389831',
+      '9955235845',
+      '9784163695808',
+      '416369580X',
+      '9780307269195',
+      '0307269477',
+      'OL382524A',
+      'Health & Fitness',
+      'Accessible book',
+      'Protected DAISY',
+      'Japanese Authors',
+      'New Age',
+      'Running',
+      'OverDrive',
+      'Nonfiction',
+      'Marathon running',
+      'Biography & Autobiography',
+      'Biography',
+      'Hashiru koto ni tsuite kataru toki ni boku no kataru koto',
+      '/works/OL2625422W',
+      'Haruki Murakami ; translated from the Japanese by Philip Gabriel.',
+      'by Haruki Murakami ; translated from the Japanese by Philip Gabriel.',
+      'Murakami Haruki.',
+      'Bungei Shunj\u016b',
+      'Alfred A. Knopf',
+      'Vintage International',
+      'baltos lankos',
+      'Knopf Doubleday Publishing Group',
+      'What I Talk About When I Talk About Running',
+      'What I talk about when I talk about running',
+      'Hashiru koto ni tsuite kataru toki ni boku no kataru koto.',
+      '2008017774',
+      '2009294820',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      '"I\'m in Kauai, in Hawaii, today, Friday, August 5, 2005."',
+      'Haruki Murakami (1949-)',
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Gabriel, Philip, 1953-', 'Gabriel, Philip.'],
+    ia_loaded_id: ['whatitalkaboutwh00mura'],
+    seed: [
+      '/books/OL25620818M',
+      '/books/OL23125654M',
+      '/books/OL24270384M',
+      '/books/OL16819451M',
+      '/books/OL24241204M',
+      '/books/OL23873280M',
+      '/books/OL24095056M',
+      '/works/OL2625422W',
+      '/subjects/marathon_running',
+      '/subjects/japanese_authors',
+      '/subjects/biography',
+      '/subjects/running',
+      '/subjects/overdrive',
+      '/subjects/biography_&_autobiography',
+      '/subjects/health_&_fitness',
+      '/subjects/new_age',
+      '/subjects/nonfiction',
+      '/subjects/protected_daisy',
+      '/subjects/person:haruki_murakami_(1949-)',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    oclc: ['246707220'],
+    ia: ['whatitalkaboutwh00mura_812', 'whatitalkaboutwh00mura'],
+    isbn: [
+      '978-9955-23-584-2',
+      '0307269191',
+      '9780307269478',
+      '0307389839',
+      '9780307389831',
+      '9955235845',
+      '9784163695808',
+      '416369580X',
+      '9780307269195',
+      '0307269477',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Health & Fitness',
+      'Accessible book',
+      'Protected DAISY',
+      'Japanese Authors',
+      'New Age',
+      'Running',
+      'OverDrive',
+      'Nonfiction',
+      'Marathon running',
+      'Biography & Autobiography',
+      'Biography',
+    ],
+    title: 'Hashiru koto ni tsuite kataru toki ni boku no kataru koto',
+    ia_collection_s: 'printdisabled;internetarchivebooks;china;librarygenesis',
+    publish_date: ['2008', '2009', '2007'],
+    type: 'work',
+    ebook_count_i: 2,
+    publish_place: [
+      'T\u014dky\u014d',
+      'New York, NY',
+      'New York',
+      'Kaunas, Lithuania',
+    ],
+    ia_box_id: ['IA106401'],
+    edition_count: 7,
+    overdrive_s: 'B59A68FA-DDEA-40B3-960C-D5C734FCB132',
+    key: '/works/OL2625422W',
+    id_goodreads: ['2164933', '2195464'],
+    public_scan_b: false,
+    id_overdrive: ['B59A68FA-DDEA-40B3-960C-D5C734FCB132'],
+    publisher: [
+      'Bungei Shunj\u016b',
+      'Alfred A. Knopf',
+      'Vintage International',
+      'baltos lankos',
+      'Knopf Doubleday Publishing Group',
+    ],
+    language: ['lit', 'eng', 'jpn'],
+    lccn: ['2008017774', '2009294820'],
+    last_modified_i: 1410781618,
+    first_publish_year: 2007,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL16819451M',
+    first_sentence: [
+      '"I\'m in Kauai, in Hawaii, today, Friday, August 5, 2005."',
+    ],
+    person: ['Haruki Murakami (1949-)'],
+    publish_year: [2008, 2009, 2007],
+    printdisabled_s: 'OL24270384M;OL16819451M',
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Sputnik Sweetheart',
+    edition_key: [
+      'OL24270277M',
+      'OL17014186M',
+      'OL7425229M',
+      'OL22375873M',
+      'OL9272821M',
+      'OL7318166M',
+      'OL7426986M',
+    ],
+    cover_i: 226274,
+    subtitle: 'a novel',
+    has_fulltext: true,
+    id_librarything: ['74326'],
+    text: [
+      'OL24270277M',
+      'OL17014186M',
+      'OL7425229M',
+      'OL22375873M',
+      'OL9272821M',
+      'OL7318166M',
+      'OL7426986M',
+      'a novel',
+      'Murakami Haruki',
+      'Gabriel, J. Philip',
+      'Gabriel, J. Philip.',
+      '294955073',
+      'sputniksweethear00mura',
+      '0375413464',
+      '9780375413469',
+      '9780375411694',
+      '9780099448471',
+      '0375411690',
+      '9780375726057',
+      '0099448475',
+      '0375726055',
+      'OL382524A',
+      'In library',
+      'Literature',
+      'Accessible book',
+      'Unrequited love',
+      'OverDrive',
+      'Teachers',
+      'Protected DAISY',
+      'Missing persons',
+      'Fiction',
+      'Women novelists',
+      'Businesswomen',
+      'Sputnik Sweetheart',
+      '/works/OL2625531W',
+      'by Haruki Murakami ; translated from the Japanese by Philip Gabriel.',
+      'by Haruki Murakami ; translated from the Japanese by Philip Gabriel',
+      'Vintage',
+      'Knopf',
+      'Alfred A Knopf',
+      'Vintage International',
+      'Knopf Doubleday Publishing Group',
+      'The sputnik sweetheart',
+      'Sup\u016btoniku no koibito.',
+      '00062004',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'In the spring of her twenty-second year, Sumire fell in love for the first time in her life.',
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Gabriel, J. Philip', 'Gabriel, J. Philip.'],
+    ia_loaded_id: ['sputniksweethear00mura'],
+    seed: [
+      '/books/OL24270277M',
+      '/books/OL17014186M',
+      '/books/OL7425229M',
+      '/books/OL22375873M',
+      '/books/OL9272821M',
+      '/books/OL7318166M',
+      '/books/OL7426986M',
+      '/works/OL2625531W',
+      '/subjects/in_library',
+      '/subjects/literature',
+      '/subjects/unrequited_love',
+      '/subjects/overdrive',
+      '/subjects/teachers',
+      '/subjects/protected_daisy',
+      '/subjects/missing_persons',
+      '/subjects/fiction',
+      '/subjects/women_novelists',
+      '/subjects/businesswomen',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    oclc: ['294955073'],
+    ia: ['sputniksweethear00mura'],
+    isbn: [
+      '0375413464',
+      '9780375413469',
+      '9780375411694',
+      '9780099448471',
+      '0375411690',
+      '9780375726057',
+      '0099448475',
+      '0375726055',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'In library',
+      'Literature',
+      'Accessible book',
+      'Unrequited love',
+      'OverDrive',
+      'Teachers',
+      'Protected DAISY',
+      'Missing persons',
+      'Fiction',
+      'Women novelists',
+      'Businesswomen',
+    ],
+    title: 'Sputnik Sweetheart',
+    lending_identifier_s: 'sputniksweethear00mura',
+    ia_collection_s:
+      'printdisabled;china;americana;internetarchivebooks;delawarecountydistrictlibrary;browserlending;inlibrary',
+    publish_date: [
+      '2002',
+      'October 3, 2002',
+      'April 9, 2002',
+      '2001',
+      'April 24, 2001',
+    ],
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['New York', 'London, United Kingdom'],
+    ia_box_id: ['IA179401'],
+    edition_count: 7,
+    first_publish_year: 2001,
+    key: '/works/OL2625531W',
+    id_goodreads: ['397165', '9557', '45316'],
+    public_scan_b: false,
+    id_overdrive: ['73907451-DAAC-43DD-8ABD-D4E0336A8CF1'],
+    publisher: [
+      'Vintage',
+      'Knopf',
+      'Alfred A Knopf',
+      'Vintage International',
+      'Knopf Doubleday Publishing Group',
+    ],
+    language: ['eng'],
+    lccn: ['00062004'],
+    last_modified_i: 1500713058,
+    lending_edition_s: 'OL7426986M',
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL7425229M',
+    first_sentence: [
+      'In the spring of her twenty-second year, Sumire fell in love for the first time in her life.',
+    ],
+    publish_year: [2002, 2001],
+    printdisabled_s: 'OL7426986M',
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'After Dark',
+    edition_key: [
+      'OL9396183M',
+      'OL9369915M',
+      'OL7915168M',
+      'OL7441786M',
+      'OL17906205M',
+      'OL24268089M',
+      'OL9644073M',
+    ],
+    cover_i: 243726,
+    id_librarything: ['637581'],
+    has_fulltext: true,
+    text: [
+      'OL9396183M',
+      'OL9369915M',
+      'OL7915168M',
+      'OL7441786M',
+      'OL17906205M',
+      'OL24268089M',
+      'OL9644073M',
+      '0739343068',
+      '0385663463',
+      '9780307265838',
+      '0307267016',
+      '0307265838',
+      '9780739343067',
+      '0676979602',
+      '9780307267016',
+      '9780676979602',
+      '9780385663465',
+      '1846550483',
+      '9781846550485',
+      'Murakami Haruki',
+      'Janet Song (Narrator)',
+      'Rubin, Jay.',
+      'Jay Rubin (Translator)',
+      '244441114',
+      '80431924',
+      '144602165',
+      'afterdark00mura',
+      'afterdark00mura_325',
+      'OL382524A',
+      'Students',
+      'Accessible book',
+      'Models (Persons)',
+      'Fiction',
+      'OverDrive',
+      'Protected DAISY',
+      'Sisters',
+      'Internet Archive Wishlist',
+      'After Dark',
+      '/works/OL2625529W',
+      'by Haruki Murakami ; translated from the Japanese by Jay Rubin.',
+      'RH Audio',
+      'Knopf',
+      'Harvill Press',
+      'Bond Street Books',
+      'Alfred A. Knopf',
+      'Vintage Canada',
+      'Knopf Doubleday Publishing Group',
+      'After dark',
+      'Afut\u0101 d\u0101ku.',
+      '2007004828',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Tokyo (Japan)',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: [
+      'Janet Song (Narrator)',
+      'Rubin, Jay.',
+      'Jay Rubin (Translator)',
+    ],
+    seed: [
+      '/books/OL9396183M',
+      '/books/OL9369915M',
+      '/books/OL7915168M',
+      '/books/OL7441786M',
+      '/books/OL17906205M',
+      '/books/OL24268089M',
+      '/books/OL9644073M',
+      '/works/OL2625529W',
+      '/subjects/students',
+      '/subjects/accessible_book',
+      '/subjects/models_(persons)',
+      '/subjects/fiction',
+      '/subjects/overdrive',
+      '/subjects/protected_daisy',
+      '/subjects/sisters',
+      '/subjects/internet_archive_wishlist',
+      '/subjects/place:tokyo_(japan)',
+      '/authors/OL382524A',
+    ],
+    oclc: ['244441114', '80431924', '144602165'],
+    ia: ['afterdark00mura', 'afterdark00mura_325'],
+    isbn: [
+      '0739343068',
+      '0385663463',
+      '9780307265838',
+      '0307267016',
+      '0307265838',
+      '9780739343067',
+      '0676979602',
+      '9780307267016',
+      '9780676979602',
+      '9780385663465',
+      '1846550483',
+      '9781846550485',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Students',
+      'Accessible book',
+      'Models (Persons)',
+      'Fiction',
+      'OverDrive',
+      'Protected DAISY',
+      'Sisters',
+      'Internet Archive Wishlist',
+    ],
+    title: 'After Dark',
+    ia_collection_s:
+      'printdisabled;internetarchivebooks;americana;librarygenesis',
+    printdisabled_s: 'OL9396183M;OL17906205M',
+    type: 'work',
+    ebook_count_i: 2,
+    publish_place: ['New York, N.Y', 'New York', 'London, United Kingdom'],
+    ia_box_id: ['IA114507'],
+    edition_count: 7,
+    key: '/works/OL2625529W',
+    id_goodreads: ['17803', '1692197', '762516', '1558894', '944624'],
+    public_scan_b: false,
+    id_overdrive: ['4494E1E3-7E12-4B9A-AC64-C0B4D2A317E2'],
+    publisher: [
+      'RH Audio',
+      'Knopf',
+      'Harvill Press',
+      'Bond Street Books',
+      'Alfred A. Knopf',
+      'Vintage Canada',
+      'Knopf Doubleday Publishing Group',
+    ],
+    language: ['eng'],
+    lccn: ['2007004828'],
+    last_modified_i: 1500027317,
+    first_publish_year: 2007,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL7441786M',
+    publish_year: [2008, 2007],
+    publish_date: [
+      'May 8, 2007',
+      'June 7, 2007',
+      'April 29, 2008',
+      'May 1, 2007',
+      '2007',
+    ],
+    place: ['Tokyo (Japan)'],
+  },
+  {
+    title_suggest: 'Blind Willow, Sleeping Woman',
+    edition_key: [
+      'OL8365296M',
+      'OL22735119M',
+      'OL9386044M',
+      'OL8364906M',
+      'OL24298756M',
+      'OL17247954M',
+    ],
+    cover_i: 5373134,
+    subtitle: '24 Stories',
+    has_fulltext: false,
+    text: [
+      'OL8365296M',
+      'OL22735119M',
+      'OL9386044M',
+      'OL8364906M',
+      'OL24298756M',
+      'OL17247954M',
+      '24 Stories',
+      'Murakami Haruki',
+      'Patrick Girard Lawlor (Editor)',
+      'Ellen Archer (Narrator)',
+      'Patrick Lawlor (Narrator)',
+      'Gabriel, J. Philip',
+      'Rubin, Jay, 1941-',
+      '77520901',
+      '71755697',
+      '72674361',
+      '177109713',
+      '9781400096084',
+      '140015295X',
+      '9781843432692',
+      '1843432692',
+      '9781400102952',
+      '1400132959',
+      '9781843433064',
+      '1400096081',
+      '0307387623',
+      '9780307387622',
+      '9781400132959',
+      '1400102952',
+      '1843433060',
+      '9781400152957',
+      'OL382524A',
+      'twenty-four stories',
+      'Blind Willow, Sleeping Woman',
+      'Haruki Murakami.',
+      'Haruki Murakami ; translated from the Japanese by Philip Gabriel and Jay Rubin',
+      'OverDrive',
+      'Literature',
+      'Translations into English',
+      'Japanese Short stories',
+      'Fiction',
+      '/works/OL2625528W',
+      'Vintage International',
+      'Knopf Doubleday Publishing Group',
+      'Tantor Media',
+      'Harvill Secker',
+      'Blind willow, sleeping woman',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Haruki Murakami (1949-)',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL8365296M',
+      '/books/OL22735119M',
+      '/books/OL9386044M',
+      '/books/OL8364906M',
+      '/books/OL24298756M',
+      '/books/OL17247954M',
+      '/works/OL2625528W',
+      '/subjects/translations_into_english',
+      '/subjects/japanese_short_stories',
+      '/subjects/overdrive',
+      '/subjects/fiction',
+      '/subjects/literature',
+      '/subjects/person:haruki_murakami_(1949-)',
+      '/authors/OL382524A',
+    ],
+    oclc: ['77520901', '71755697', '72674361', '177109713'],
+    contributor: [
+      'Patrick Girard Lawlor (Editor)',
+      'Ellen Archer (Narrator)',
+      'Patrick Lawlor (Narrator)',
+      'Gabriel, J. Philip',
+      'Rubin, Jay, 1941-',
+    ],
+    isbn: [
+      '9781400096084',
+      '140015295X',
+      '9781843432692',
+      '1843432692',
+      '9781400102952',
+      '1400132959',
+      '9781843433064',
+      '1400096081',
+      '0307387623',
+      '9780307387622',
+      '9781400132959',
+      '1400102952',
+      '1843433060',
+      '9781400152957',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'OverDrive',
+      'Literature',
+      'Translations into English',
+      'Japanese Short stories',
+      'Fiction',
+    ],
+    title: 'Blind Willow, Sleeping Woman',
+    publish_date: ['October 1, 2006', '2006', '2007'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['New York', 'London'],
+    id_librarything: ['748397'],
+    edition_count: 6,
+    overdrive_s: 'C17C2BC6-D6BC-4535-819A-F1F8866E52D9',
+    key: '/works/OL2625528W',
+    id_goodreads: ['964388', '1205453', '818122', '256574', '9832', '45296'],
+    id_overdrive: ['C17C2BC6-D6BC-4535-819A-F1F8866E52D9'],
+    publisher: [
+      'Vintage International',
+      'Knopf Doubleday Publishing Group',
+      'Tantor Media',
+      'Harvill Secker',
+    ],
+    language: ['eng'],
+    last_modified_i: 1375625901,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL17247954M',
+    person: ['Haruki Murakami (1949-)'],
+    publish_year: [2006, 2007],
+    first_publish_year: 2006,
+  },
+  {
+    title_suggest: 'The elephant vanishes',
+    edition_key: [
+      'OL17218963M',
+      'OL1744979M',
+      'OL1431988M',
+      'OL21007350M',
+      'OL7318180M',
+      'OL9272835M',
+    ],
+    cover_i: 4009808,
+    id_librarything: ['2940'],
+    has_fulltext: true,
+    text: [
+      'OL17218963M',
+      'OL1744979M',
+      'OL1431988M',
+      'OL21007350M',
+      'OL7318180M',
+      'OL9272835M',
+      '9780099448754',
+      '0679420576',
+      '9780679750536',
+      '9780679420576',
+      '9780241001653',
+      '0679750533',
+      '024100165X',
+      '0099448750',
+      '9789621713919',
+      '9621713919',
+      'Murakami Haruki',
+      'Birnbaum, Alfred.',
+      'Rubin, Jay.',
+      '59890352',
+      '32372879',
+      'elephantvanishes00mura',
+      'OL382524A',
+      'stories',
+      'The elephant vanishes',
+      'Social life and customs',
+      'Humorous stories',
+      'Accessible book',
+      'Translations into English',
+      'Fiction',
+      'Protected DAISY',
+      '/works/OL2625530W',
+      'stories by Haruki Murakami ; translated from the Japanese by Alfred Birnbaum and Jay Rubin.',
+      'by Haruki Murakami ; translated from the Japanese by Alfred Birnbaum and Jay Rubin.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Vintage',
+      'Vintage Books',
+      'A. A. Knopf',
+      'Bo Yi',
+      'H. Hamilton',
+      'Distributed by Random House',
+      'The Elephant vanishes',
+      'The Elephant Vanishes',
+      'The elephant vanishes.',
+      '92054277',
+      '93043493',
+      'Haruki Murakami (1949-)',
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Birnbaum, Alfred.', 'Rubin, Jay.'],
+    seed: [
+      '/books/OL17218963M',
+      '/books/OL1744979M',
+      '/books/OL1431988M',
+      '/books/OL21007350M',
+      '/books/OL7318180M',
+      '/books/OL9272835M',
+      '/works/OL2625530W',
+      '/subjects/social_life_and_customs',
+      '/subjects/translations_into_english',
+      '/subjects/fiction',
+      '/subjects/humorous_stories',
+      '/subjects/protected_daisy',
+      '/subjects/person:haruki_murakami_(1949-)',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    oclc: ['59890352', '32372879'],
+    ia: ['elephantvanishes00mura'],
+    isbn: [
+      '9780099448754',
+      '0679420576',
+      '9780679750536',
+      '9780679420576',
+      '9780241001653',
+      '0679750533',
+      '024100165X',
+      '0099448750',
+      '9789621713919',
+      '9621713919',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Social life and customs',
+      'Humorous stories',
+      'Accessible book',
+      'Translations into English',
+      'Fiction',
+      'Protected DAISY',
+    ],
+    title: 'The elephant vanishes',
+    ia_collection_s: 'printdisabled;americana;internetarchivebooks',
+    publish_date: ['February 8, 2001', '1993', '1995', '1994'],
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['Hong Kong', 'New York', 'London'],
+    printdisabled_s: 'OL1744979M',
+    ia_box_id: ['IA138114'],
+    edition_count: 6,
+    key: '/works/OL2625530W',
+    id_goodreads: ['9556', '9555', '1022990', '45307'],
+    public_scan_b: false,
+    publisher: [
+      'Vintage',
+      'Vintage Books',
+      'A. A. Knopf',
+      'Bo Yi',
+      'H. Hamilton',
+      'Distributed by Random House',
+    ],
+    language: ['chi', 'eng'],
+    lccn: ['92054277', '93043493'],
+    last_modified_i: 1320446437,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL1744979M',
+    person: ['Haruki Murakami (1949-)'],
+    publish_year: [1993, 2001, 1995, 1994],
+    first_publish_year: 1993,
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Norwegian Wood',
+    edition_key: [
+      'OL14661734M',
+      'OL7426270M',
+      'OL9318947M',
+      'OL8626298M',
+      'OL24608213M',
+      'OL25837363M',
+    ],
+    isbn: [
+      '9789985332047',
+      '0870119435',
+      '1860468187',
+      '0099554569',
+      '0375704027',
+      '9781417709069',
+      '1417709065',
+      '4061860526',
+      '9780870119439',
+      '9784061860520',
+      '9780099554561',
+      '9985332040',
+      '9781860468186',
+      '9780375704024',
+      '4061860518',
+      '9784061860513',
+    ],
+    has_fulltext: true,
+    text: [
+      'OL14661734M',
+      'OL7426270M',
+      'OL9318947M',
+      'OL8626298M',
+      'OL24608213M',
+      'OL25837363M',
+      '9789985332047',
+      '0870119435',
+      '1860468187',
+      '0099554569',
+      '0375704027',
+      '9781417709069',
+      '1417709065',
+      '4061860526',
+      '9780870119439',
+      '9784061860520',
+      '9780099554561',
+      '9985332040',
+      '9781860468186',
+      '9780375704024',
+      '4061860518',
+      '9784061860513',
+      'Murakami Haruki',
+      'Jay Rubin (Translator)',
+      'norwegianwoodvol01mura',
+      'norwegianwoodpan00mura',
+      'OL382524A',
+      'College students',
+      'Accessible book',
+      'Protected DAISY',
+      'Fiction',
+      'Norwegian Wood',
+      '/works/OL15338505W',
+      'Haruki Murakami ; translated by Alfred Birnbaum',
+      'Tandem Library',
+      'Kodansha International',
+      'Vintage Books',
+      'Varrak',
+      'Vintage',
+      'The Harvill Press',
+      'Norra mets',
+      'Norwegian wood',
+      'Noruwei no mori.',
+      '89024554',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'I WAS THIRTY-SEVEN THEN, STRAPPED IN MY SEAT AS THE HUGE 747 plunged through dense cloud cover on approach to the Hamburg airport.',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Jay Rubin (Translator)'],
+    seed: [
+      '/books/OL14661734M',
+      '/books/OL7426270M',
+      '/books/OL9318947M',
+      '/books/OL8626298M',
+      '/books/OL24608213M',
+      '/books/OL25837363M',
+      '/works/OL15338505W',
+      '/subjects/college_students',
+      '/subjects/fiction',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['5860', '8600632'],
+    ia: ['norwegianwoodvol01mura', 'norwegianwoodpan00mura'],
+    author_key: ['OL382524A'],
+    subject: [
+      'College students',
+      'Accessible book',
+      'Protected DAISY',
+      'Fiction',
+    ],
+    title: 'Norwegian Wood',
+    ia_collection_s: 'printdisabled;librarygenesis',
+    printdisabled_s: 'OL14661734M;OL8626298M',
+    type: 'work',
+    ebook_count_i: 2,
+    publish_place: ['Great Britain', 'New York', 'Tallinn, Estonia', 'Tokyo'],
+    edition_count: 6,
+    key: '/works/OL15338505W',
+    id_goodreads: [
+      '24858736',
+      '234716',
+      '864894',
+      '1712090',
+      '1712091',
+      '11297',
+      '3194869',
+    ],
+    public_scan_b: false,
+    publisher: [
+      'Tandem Library',
+      'Kodansha International',
+      'Vintage Books',
+      'Varrak',
+      'Vintage',
+      'The Harvill Press',
+    ],
+    language: ['est', 'eng'],
+    lccn: ['89024554'],
+    last_modified_i: 1484404315,
+    first_publish_year: 1989,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL7426270M',
+    first_sentence: [
+      'I WAS THIRTY-SEVEN THEN, STRAPPED IN MY SEAT AS THE HUGE 747 plunged through dense cloud cover on approach to the Hamburg airport.',
+    ],
+    publish_year: [2015, 2003, 2000, 2001, 1989],
+    publish_date: [
+      'September 12, 2000',
+      'September 2000',
+      '1989',
+      '2003',
+      '2001',
+      '2015',
+    ],
+  },
+  {
+    title_suggest: 'Underground',
+    edition_key: [
+      'OL7426974M',
+      'OL18331553M',
+      'OL22036994M',
+      'OL9360344M',
+      'OL9273123M',
+      'OL7318468M',
+    ],
+    cover_i: 75473,
+    subtitle: 'The Tokyo Gas Attack and the Japanese Psyche',
+    has_fulltext: true,
+    text: [
+      'OL7426974M',
+      'OL18331553M',
+      'OL22036994M',
+      'OL9360344M',
+      'OL9273123M',
+      'OL7318468M',
+      'The Tokyo Gas Attack and the Japanese Psyche',
+      'Murakami Haruki',
+      'Philip Gabriel (Translator)',
+      'Gabriel, J. Philip.',
+      'Birnbaum, Alfred.',
+      'J. Philip Gabriel (Translator)',
+      'Alfred Birnbaum (Translator)',
+      'undergroundtokyo00haru',
+      '9781417709526',
+      '9780099461098',
+      '1417709529',
+      '0099461099',
+      '9780375725807',
+      '9781860468438',
+      '1860468438',
+      '0375725806',
+      'OL382524A',
+      'Cults',
+      'Accessible book',
+      'Victims of terrorism',
+      'Oumu Shinriky\u014d (Religious organization)',
+      'Protected DAISY',
+      'Terrorism',
+      'Internet Archive Wishlist',
+      'the Tokyo gas attack and the Japanese psyche',
+      'Underground',
+      '/works/OL2625532W',
+      'Haruki Murakami  ; translated from the Japanese by Alfred Birnbaum and Philip Gabriel.',
+      'Haruki Murakami ; translated from the Japanese by Alfred Birnbaum and Philip Gabriel.',
+      'Vintage',
+      'Vintage International',
+      'Tandem Library',
+      'Panther',
+      'Tokyo gas attack and the Japanese psyche',
+      'And\u0101guraundo.',
+      '00069310',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Two men were assigned to drop sarin gas on the Chiyoda Line: Ikuo Hayashi and Tomomitsu Niimi.',
+      'Japan',
+      'Tokyo',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: [
+      'Philip Gabriel (Translator)',
+      'Gabriel, J. Philip.',
+      'Birnbaum, Alfred.',
+      'J. Philip Gabriel (Translator)',
+      'Alfred Birnbaum (Translator)',
+    ],
+    seed: [
+      '/books/OL7426974M',
+      '/books/OL18331553M',
+      '/books/OL22036994M',
+      '/books/OL9360344M',
+      '/books/OL9273123M',
+      '/books/OL7318468M',
+      '/works/OL2625532W',
+      '/subjects/terrorism',
+      '/subjects/internet_archive_wishlist',
+      '/subjects/cults',
+      '/subjects/oumu_shinriky\u014d_(religious_organization)',
+      '/subjects/victims_of_terrorism',
+      '/subjects/place:japan',
+      '/subjects/place:tokyo',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['58240'],
+    ia: ['undergroundtokyo00haru'],
+    isbn: [
+      '9781417709526',
+      '9780099461098',
+      '1417709529',
+      '0099461099',
+      '9780375725807',
+      '9781860468438',
+      '1860468438',
+      '0375725806',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Cults',
+      'Accessible book',
+      'Victims of terrorism',
+      'Oumu Shinriky\u014d (Religious organization)',
+      'Protected DAISY',
+      'Terrorism',
+      'Internet Archive Wishlist',
+    ],
+    title: 'Underground',
+    lending_identifier_s: 'undergroundtokyo00haru',
+    ia_collection_s:
+      'printdisabled;china;americana;internetarchivebooks;delawarecountydistrictlibrary;browserlending;inlibrary',
+    printdisabled_s: 'OL18331553M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['New York', 'London'],
+    ia_box_id: ['IA1127319'],
+    edition_count: 6,
+    first_publish_year: 2001,
+    key: '/works/OL2625532W',
+    id_goodreads: ['229173', '17801', '931071', '4500624'],
+    public_scan_b: false,
+    publisher: [
+      'Vintage',
+      'Vintage International',
+      'Tandem Library',
+      'Panther',
+    ],
+    language: ['eng'],
+    lccn: ['00069310'],
+    last_modified_i: 1500067726,
+    lending_edition_s: 'OL18331553M',
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL9273123M',
+    first_sentence: [
+      'Two men were assigned to drop sarin gas on the Chiyoda Line: Ikuo Hayashi and Tomomitsu Niimi.',
+    ],
+    publish_year: [2003, 2001],
+    publish_date: ['April 2001', '2001', 'April 10, 2001', 'September 4, 2003'],
+    place: ['Japan', 'Tokyo'],
+  },
+  {
+    title_suggest: 'Kafka am strand',
+    edition_key: ['OL25554921M'],
+    isbn: ['9783832178666', '383217866X'],
+    has_fulltext: true,
+    text: [
+      'OL25554921M',
+      '9783832178666',
+      '383217866X',
+      'Haruki Murakami',
+      'Gra fe, Ursula, 1956-',
+      '56112515',
+      'kafkaamstrand00mura',
+      'OL7087900A',
+      'Kafka am strand',
+      'Runaway teenagers',
+      'Teenage boys',
+      'Protected DAISY',
+      'Accessible book',
+      'Fiction',
+      '/works/OL16960980W',
+      'Haruki Murakami ; aus dem Japanischen von Ursula Gra fe',
+      'DuMont',
+      'Umibe no Kafuka',
+      'Japan',
+    ],
+    author_name: ['Haruki Murakami'],
+    ia: ['kafkaamstrand00mura'],
+    seed: [
+      '/books/OL25554921M',
+      '/works/OL16960980W',
+      '/subjects/runaway_teenagers',
+      '/subjects/teenage_boys',
+      '/subjects/fiction',
+      '/subjects/place:japan',
+      '/authors/OL7087900A',
+    ],
+    oclc: ['56112515'],
+    contributor: ['Gra fe, Ursula, 1956-'],
+    author_key: ['OL7087900A'],
+    subject: [
+      'Runaway teenagers',
+      'Teenage boys',
+      'Protected DAISY',
+      'Accessible book',
+      'Fiction',
+    ],
+    title: 'Kafka am strand',
+    ia_collection_s: 'printdisabled;librarygenesis',
+    first_publish_year: 2004,
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['Ko ln'],
+    printdisabled_s: 'OL25554921M',
+    edition_count: 1,
+    key: '/works/OL16960980W',
+    public_scan_b: false,
+    publisher: ['DuMont'],
+    language: ['ger'],
+    last_modified_i: 1406632390,
+    publish_year: [2004],
+    publish_date: ['2004'],
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Tony Takitani',
+    edition_key: ['OL25542816M'],
+    isbn: ['3832179356', '9783832179359'],
+    has_fulltext: true,
+    text: [
+      'OL25542816M',
+      '3832179356',
+      '9783832179359',
+      'Haruki Murakami',
+      '76742954',
+      'tonytakitani00mura',
+      'OL7203041A',
+      'Tony Takitani',
+      'Ehefrau',
+      'Kleidung',
+      'Zeichner',
+      'Sucht',
+      'Japanisch',
+      'Accessible book',
+      'Erza hlung',
+      'Belletristische Darstellung',
+      'Protected DAISY',
+      '/works/OL16934995W',
+      'Haruki Murakami. Aus dem Japan. von Ursula Gra fe',
+      'DuMont',
+      'Japan',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL25542816M',
+      '/works/OL16934995W',
+      '/subjects/ehefrau',
+      '/subjects/belletristische_darstellung',
+      '/subjects/kleidung',
+      '/subjects/zeichner',
+      '/subjects/sucht',
+      '/subjects/erza_hlung',
+      '/subjects/japanisch',
+      '/subjects/place:japan',
+      '/authors/OL7203041A',
+    ],
+    oclc: ['76742954'],
+    ia: ['tonytakitani00mura'],
+    author_key: ['OL7203041A'],
+    subject: [
+      'Ehefrau',
+      'Kleidung',
+      'Zeichner',
+      'Sucht',
+      'Japanisch',
+      'Accessible book',
+      'Erza hlung',
+      'Belletristische Darstellung',
+      'Protected DAISY',
+    ],
+    title: 'Tony Takitani',
+    ia_collection_s: 'printdisabled;librarygenesis',
+    first_publish_year: 2005,
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['Ko ln'],
+    printdisabled_s: 'OL25542816M',
+    edition_count: 1,
+    key: '/works/OL16934995W',
+    public_scan_b: false,
+    publisher: ['DuMont'],
+    language: ['ger'],
+    last_modified_i: 1406577896,
+    publish_year: [2005],
+    publish_date: ['2005'],
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Sputnik Sweetheart',
+    edition_key: ['OL25550378M'],
+    isbn: ['9783832156961', '3832156968'],
+    has_fulltext: true,
+    text: [
+      'OL25550378M',
+      '9783832156961',
+      '3832156968',
+      'Haruki Murakami',
+      'Gra fe, Ursula',
+      '255770886',
+      'sputniksweethear00mura_651',
+      'OL7203041A',
+      'Roman',
+      'Sputnik Sweetheart',
+      'Haruki Murakami. Aus dem Japan. von Ursula Gra fe',
+      'Bindungslosigkeit',
+      'Gescha ftsfrau',
+      'Schriftstellerin',
+      'Accessible book',
+      'Lesbische Liebe',
+      'Protected DAISY',
+      '/works/OL16951001W',
+      'Supu toniku-no-koibito',
+      'DuMont',
+    ],
+    author_name: ['Haruki Murakami'],
+    ia: ['sputniksweethear00mura_651'],
+    seed: [
+      '/books/OL25550378M',
+      '/works/OL16951001W',
+      '/subjects/lesbische_liebe',
+      '/subjects/bindungslosigkeit',
+      '/subjects/gescha_ftsfrau',
+      '/subjects/schriftstellerin',
+      '/authors/OL7203041A',
+    ],
+    oclc: ['255770886'],
+    contributor: ['Gra fe, Ursula'],
+    author_key: ['OL7203041A'],
+    subject: [
+      'Bindungslosigkeit',
+      'Gescha ftsfrau',
+      'Schriftstellerin',
+      'Accessible book',
+      'Lesbische Liebe',
+      'Protected DAISY',
+    ],
+    title: 'Sputnik Sweetheart',
+    ia_collection_s: 'printdisabled;librarygenesis',
+    publish_date: ['2003'],
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['Ko ln'],
+    printdisabled_s: 'OL25550378M',
+    edition_count: 1,
+    key: '/works/OL16951001W',
+    public_scan_b: false,
+    publisher: ['DuMont'],
+    language: ['ger'],
+    last_modified_i: 1406608970,
+    publish_year: [2003],
+    first_publish_year: 2003,
+  },
+  {
+    title_suggest: 'Taira no Masakado densetsu',
+    edition_key: ['OL19182841M'],
+    isbn: ['4762941611', '9784762941610'],
+    has_fulltext: false,
+    text: [
+      'OL19182841M',
+      '4762941611',
+      '9784762941610',
+      'Haruki Murakami',
+      '48374560',
+      'OL5036666A',
+      'Taira no Masakado densetsu',
+      'Murakami Haruki cho.',
+      'Generals',
+      'Legends',
+      'Biography',
+      '/works/OL11808285W',
+      'Ky\u016bko Shoin',
+      'Masakado Taira (903 or 4-940)',
+      'Japan',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL19182841M',
+      '/works/OL11808285W',
+      '/subjects/generals',
+      '/subjects/legends',
+      '/subjects/biography',
+      '/subjects/person:masakado_taira_(903_or_4-940)',
+      '/subjects/place:japan',
+      '/authors/OL5036666A',
+    ],
+    oclc: ['48374560'],
+    author_key: ['OL5036666A'],
+    subject: ['Generals', 'Legends', 'Biography'],
+    title: 'Taira no Masakado densetsu',
+    publish_date: ['2001'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL11808285W',
+    id_goodreads: ['2376602'],
+    publisher: ['Ky\u016bko Shoin'],
+    language: ['jpn'],
+    last_modified_i: 1303787784,
+    person: ['Masakado Taira (903 or 4-940)'],
+    publish_year: [2001],
+    first_publish_year: 2001,
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Taira no Masakado',
+    edition_key: ['OL16484227M'],
+    isbn: ['9784762942006', '4762942006'],
+    has_fulltext: false,
+    text: [
+      'OL16484227M',
+      '9784762942006',
+      '4762942006',
+      'Haruki Murakami',
+      '183618126',
+      'OL5036666A',
+      'ch\u014dsa to kenky\u016b',
+      'Taira no Masakado',
+      'Murakami Haruki cho.',
+      'Generals',
+      'Bibliography',
+      'Biography',
+      '/works/OL11808284W',
+      'Ky\u016bko Shoin',
+      'Taira Masakado',
+      'Sh\u014dmon ki.',
+      '2007430709',
+      'Masakado Taira (903 or 4-940)',
+      'Japan',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL16484227M',
+      '/works/OL11808284W',
+      '/subjects/generals',
+      '/subjects/bibliography',
+      '/subjects/biography',
+      '/subjects/person:masakado_taira_(903_or_4-940)',
+      '/subjects/place:japan',
+      '/authors/OL5036666A',
+    ],
+    oclc: ['183618126'],
+    author_key: ['OL5036666A'],
+    subject: ['Generals', 'Bibliography', 'Biography'],
+    title: 'Taira no Masakado',
+    publish_date: ['1937'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL11808284W',
+    id_goodreads: ['2376588'],
+    publisher: ['Ky\u016bko Shoin'],
+    language: ['jpn'],
+    lccn: ['2007430709'],
+    last_modified_i: 1307515140,
+    person: ['Masakado Taira (903 or 4-940)'],
+    publish_year: [1937],
+    first_publish_year: 1937,
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Shinpukuji-bon, Y\u014d Shukei-bon Sh\u014dmon ki shinkai',
+    edition_key: ['OL19355265M'],
+    isbn: ['9784762935138', '4762935131'],
+    has_fulltext: false,
+    text: [
+      'OL19355265M',
+      '9784762935138',
+      '4762935131',
+      'Haruki Murakami',
+      '56119869',
+      'OL5036666A',
+      'Shinpukuji-bon, Y\u014d Shukei-bon Sh\u014dmon ki shinkai',
+      'Murakami Haruki cho.',
+      'Sh\u014dmonki',
+      'Sh\u014dmon ki',
+      '/works/OL11808283W',
+      'Ky\u016bko Shoin',
+      'Sh\u014dmon ki',
+      'Masakado Taira (903 or 4-940)',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL19355265M',
+      '/works/OL11808283W',
+      '/subjects/sh\u014dmon_ki',
+      '/subjects/sh\u014dmonki',
+      '/subjects/person:masakado_taira_(903_or_4-940)',
+      '/authors/OL5036666A',
+    ],
+    oclc: ['56119869'],
+    author_key: ['OL5036666A'],
+    subject: ['Sh\u014dmonki', 'Sh\u014dmon ki'],
+    title: 'Shinpukuji-bon, Y\u014d Shukei-bon Sh\u014dmon ki shinkai',
+    publish_date: ['2004'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL11808283W',
+    id_goodreads: ['2376600'],
+    publisher: ['Ky\u016bko Shoin'],
+    language: ['jpn'],
+    last_modified_i: 1304126320,
+    person: ['Masakado Taira (903 or 4-940)'],
+    publish_year: [2004],
+    first_publish_year: 2004,
+  },
+  {
+    title_suggest: 'Underground',
+    publish_place: ['Barcelona, Spain'],
+    isbn: ['8483839490', '978-84-8383-949-2'],
+    has_fulltext: false,
+    edition_count: 1,
+    edition_key: ['OL25710484M'],
+    last_modified_i: 1435922924,
+    title: 'Underground',
+    publisher: ['Tusquets'],
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL25710484M',
+    seed: ['/books/OL25710484M', '/works/OL17139878W', '/authors/OL5036666A'],
+    first_publish_year: 2014,
+    publish_year: [2014],
+    key: '/works/OL17139878W',
+    text: [
+      'OL25710484M',
+      '8483839490',
+      '978-84-8383-949-2',
+      'Haruki Murakami',
+      'OL5036666A',
+      'El atentado con gas sar\u00edn en el metro de Tokio y la psicolog\u00eda japonesa',
+      'Underground',
+      '/works/OL17139878W',
+      'traducci\u00f3n de Fernando Cordob\u00e9s y Yoko Ogihara',
+      'Tusquets',
+    ],
+    publish_date: ['2014'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'After dark',
+    edition_key: ['OL25537207M'],
+    isbn: ['9788483835623', '8483835622'],
+    has_fulltext: true,
+    text: [
+      'OL25537207M',
+      '9788483835623',
+      '8483835622',
+      'Haruki Murakami',
+      'Porta Fuentes, Lourdes',
+      '676789335',
+      'afterdark00mura_046',
+      'OL7087900A',
+      'Modelos',
+      'Students',
+      'Accessible book',
+      'Hermanas',
+      'Models (Persons)',
+      'Fiction',
+      'Estudiantes',
+      'Novela',
+      'Protected DAISY',
+      'Sisters',
+      'After dark',
+      '/works/OL16921333W',
+      'Haruki Murakami ; traduccio n del japone s de Lourdes Porta',
+      'Tusquets Editores',
+      'Afuta  da ku',
+      'Tokyo (Japan)',
+      'Tokyo (Japo n)',
+    ],
+    author_name: ['Haruki Murakami'],
+    ia: ['afterdark00mura_046'],
+    seed: [
+      '/books/OL25537207M',
+      '/works/OL16921333W',
+      '/subjects/modelos',
+      '/subjects/students',
+      '/subjects/hermanas',
+      '/subjects/models_(persons)',
+      '/subjects/fiction',
+      '/subjects/estudiantes',
+      '/subjects/novela',
+      '/subjects/sisters',
+      '/subjects/place:tokyo_(japan)',
+      '/subjects/place:tokyo_(japo_n)',
+      '/authors/OL7087900A',
+    ],
+    oclc: ['676789335'],
+    contributor: ['Porta Fuentes, Lourdes'],
+    author_key: ['OL7087900A'],
+    subject: [
+      'Modelos',
+      'Students',
+      'Accessible book',
+      'Hermanas',
+      'Models (Persons)',
+      'Fiction',
+      'Estudiantes',
+      'Novela',
+      'Protected DAISY',
+      'Sisters',
+    ],
+    title: 'After dark',
+    ia_collection_s: 'printdisabled;librarygenesis',
+    printdisabled_s: 'OL25537207M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['Barcelona'],
+    edition_count: 1,
+    key: '/works/OL16921333W',
+    public_scan_b: false,
+    publisher: ['Tusquets Editores'],
+    language: ['spa'],
+    last_modified_i: 1406701390,
+    first_publish_year: 2010,
+    publish_year: [2010],
+    publish_date: ['2010'],
+    place: ['Tokyo (Japan)', 'Tokyo (Japo n)'],
+  },
+  {
+    title_suggest: 'Cr\u00f3nica del p\u00e1jaro que da cuerda al mundo',
+    publish_place: ['Barcelona, Spain'],
+    isbn: ['848383510X', '978-84-8383-510-4'],
+    has_fulltext: false,
+    language: ['spa'],
+    edition_count: 1,
+    edition_key: ['OL25438123M'],
+    last_modified_i: 1392633035,
+    title: 'Cr\u00f3nica del p\u00e1jaro que da cuerda al mundo',
+    publisher: ['Tusquets'],
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL25438123M',
+    seed: ['/books/OL25438123M', '/works/OL16812561W', '/authors/OL7087900A'],
+    first_publish_year: 2012,
+    publish_year: [2012],
+    key: '/works/OL16812561W',
+    text: [
+      'OL25438123M',
+      '848383510X',
+      '978-84-8383-510-4',
+      'Haruki Murakami',
+      'OL7087900A',
+      'Cr\u00f3nica del p\u00e1jaro que da cuerda al mundo',
+      '/works/OL16812561W',
+      'Tusquets',
+    ],
+    publish_date: ['2012'],
+    author_key: ['OL7087900A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'Hombres sin mujeres',
+    publisher: ['Tusquets'],
+    isbn: ['978-84-9066-043-0', '8490660433'],
+    has_fulltext: false,
+    edition_count: 1,
+    edition_key: ['OL25662665M'],
+    last_modified_i: 1426607263,
+    title: 'Hombres sin mujeres',
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL25662665M',
+    seed: ['/books/OL25662665M', '/works/OL17092488W', '/authors/OL5036666A'],
+    first_publish_year: 2015,
+    publish_year: [2015],
+    key: '/works/OL17092488W',
+    text: [
+      'OL25662665M',
+      '978-84-9066-043-0',
+      '8490660433',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Hombres sin mujeres',
+      '/works/OL17092488W',
+      'Tusquets',
+    ],
+    publish_date: ['2015'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: '19Q4',
+    publisher: ['Vintage'],
+    isbn: ['0307476464', '978-0307476463'],
+    has_fulltext: false,
+    edition_count: 1,
+    edition_key: ['OL25796166M'],
+    last_modified_i: 1446539928,
+    title: '19Q4',
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL25796166M',
+    seed: ['/books/OL25796166M', '/works/OL17223444W', '/authors/OL5036666A'],
+    first_publish_year: 2013,
+    publish_year: [2013],
+    key: '/works/OL17223444W',
+    text: [
+      'OL25796166M',
+      '0307476464',
+      '978-0307476463',
+      'Haruki Murakami',
+      'OL5036666A',
+      '19Q4',
+      '/works/OL17223444W',
+      'Vintage',
+    ],
+    publish_date: ['2013'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'La biblioteca secreta',
+    publisher: ['Libros del zorro rojo'],
+    isbn: ['978-84-942918-0-7', '8494291807'],
+    has_fulltext: false,
+    edition_count: 1,
+    edition_key: ['OL25888079M'],
+    last_modified_i: 1453287084,
+    title: 'La biblioteca secreta',
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL25888079M',
+    seed: ['/books/OL25888079M', '/works/OL17314101W', '/authors/OL5036666A'],
+    first_publish_year: 2014,
+    publish_year: [2014],
+    key: '/works/OL17314101W',
+    text: [
+      'OL25888079M',
+      '978-84-942918-0-7',
+      '8494291807',
+      'Haruki Murakami',
+      'OL5036666A',
+      'La biblioteca secreta',
+      '/works/OL17314101W',
+      'Libros del zorro rojo',
+    ],
+    publish_date: ['2014'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: '1Q84',
+    publish_place: ['Barcelona, Espa\u00f1a'],
+    isbn: ['848383314X', '978-84-8383-314-8'],
+    has_fulltext: false,
+    edition_count: 1,
+    edition_key: ['OL25893576M'],
+    last_modified_i: 1456402371,
+    title: '1Q84',
+    publisher: ['Tusquets'],
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL25893576M',
+    seed: ['/books/OL25893576M', '/works/OL17318775W', '/authors/OL5036666A'],
+    first_publish_year: 2011,
+    publish_year: [2011],
+    key: '/works/OL17318775W',
+    text: [
+      'OL25893576M',
+      '848383314X',
+      '978-84-8383-314-8',
+      'Haruki Murakami',
+      'OL5036666A',
+      '1Q84',
+      '/works/OL17318775W',
+      'Tusquets',
+    ],
+    publish_date: ['2011'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: '1Q84',
+    edition_key: ['OL25224595M'],
+    isbn: ['1846554071', '9781846554070'],
+    has_fulltext: false,
+    text: [
+      'OL25224595M',
+      '1846554071',
+      '9781846554070',
+      'Haruki Murakami',
+      'Gabriel, Philip, 1953-',
+      'Rubin, Jay, 1941-',
+      'OL7087900A',
+      'fantasy',
+      'magic',
+      'alternate universe',
+      '1Q84',
+      '/works/OL16533021W',
+      'Haruki Murakami ; books one and two translated from the Japanese by Jay Rubin',
+      'Harvill Secker',
+      '2012371492',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL25224595M',
+      '/works/OL16533021W',
+      '/subjects/magic',
+      '/subjects/fantasy',
+      '/subjects/alternate_universe',
+      '/authors/OL7087900A',
+    ],
+    contributor: ['Gabriel, Philip, 1953-', 'Rubin, Jay, 1941-'],
+    author_key: ['OL7087900A'],
+    subject: ['fantasy', 'magic', 'alternate universe'],
+    title: '1Q84',
+    publish_date: ['2011'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['London'],
+    edition_count: 1,
+    key: '/works/OL16533021W',
+    publisher: ['Harvill Secker'],
+    language: ['eng'],
+    lccn: ['2012371492'],
+    last_modified_i: 1496978577,
+    cover_edition_key: 'OL25224595M',
+    publish_year: [2011],
+    first_publish_year: 2011,
+  },
+  {
+    title_suggest: 'Umibe no Kafuka',
+    edition_key: ['OL26336340M'],
+    cover_i: 8016430,
+    isbn: ['4101001553', '4101001545', '9784101001555', '9784101001548'],
+    has_fulltext: true,
+    text: [
+      'OL26336340M',
+      '4101001553',
+      '4101001545',
+      '9784101001555',
+      '9784101001548',
+      'Haruki Murakami',
+      '61207074',
+      'kafkaonshorevol100haru',
+      'OL7087900A',
+      'Runaway teenagers',
+      'Detective and mystery stories',
+      'Japanese fiction',
+      'Accessible book',
+      'Fiction',
+      'Parent and child',
+      'Protected DAISY',
+      'Umibe no Kafuka',
+      '/works/OL17736913W',
+      'Murakami Haruki cho',
+      'Shincho sha',
+      'Japan',
+      '20th century',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL26336340M',
+      '/works/OL17736913W',
+      '/subjects/runaway_teenagers',
+      '/subjects/parent_and_child',
+      '/subjects/detective_and_mystery_stories',
+      '/subjects/fiction',
+      '/subjects/japanese_fiction',
+      '/subjects/place:japan',
+      '/subjects/time:20th_century',
+      '/authors/OL7087900A',
+    ],
+    oclc: ['61207074'],
+    ia: ['kafkaonshorevol100haru'],
+    author_key: ['OL7087900A'],
+    subject: [
+      'Runaway teenagers',
+      'Detective and mystery stories',
+      'Japanese fiction',
+      'Accessible book',
+      'Fiction',
+      'Parent and child',
+      'Protected DAISY',
+    ],
+    title: 'Umibe no Kafuka',
+    lending_identifier_s: 'kafkaonshorevol100haru',
+    ia_collection_s:
+      'printdisabled;china;americana;internetarchivebooks;delawarecountydistrictlibrary;browserlending;inlibrary',
+    printdisabled_s: 'OL26336340M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['To kyo'],
+    ia_box_id: ['IA155024'],
+    edition_count: 1,
+    first_publish_year: 2005,
+    key: '/works/OL17736913W',
+    public_scan_b: false,
+    publisher: ['Shincho sha'],
+    language: ['jpn'],
+    last_modified_i: 1498302499,
+    lending_edition_s: 'OL26336340M',
+    cover_edition_key: 'OL26336340M',
+    publish_year: [2005],
+    publish_date: ['2005'],
+    place: ['Japan'],
+    time: ['20th century'],
+  },
+  {
+    title_suggest: 'Nich \u00a3i \u0141a na karuseli',
+    edition_key: ['OL26349201M'],
+    cover_i: 8031601,
+    isbn: ['5699333312', '9785699333318'],
+    has_fulltext: true,
+    text: [
+      'OL26349201M',
+      '5699333312',
+      '9785699333318',
+      'Haruki Murakami',
+      'Murakami, Haruki, 1949-',
+      '317125756',
+      'nichianakaruseli00haru',
+      'OL7087900A',
+      'Accessible book',
+      'Protected DAISY',
+      'Nich \u00a3i \u0141a na karuseli',
+      '/works/OL17751318W',
+      'Kharuki Murakami ; [perevod s i \u0141aponskogo I \u0141Ulii Chanarevoi ].',
+      'E ksmo',
+      'Kaiten mokuba no deddo hi to',
+    ],
+    author_name: ['Haruki Murakami'],
+    contributor: ['Murakami, Haruki, 1949-'],
+    seed: ['/books/OL26349201M', '/works/OL17751318W', '/authors/OL7087900A'],
+    oclc: ['317125756'],
+    ia: ['nichianakaruseli00haru'],
+    author_key: ['OL7087900A'],
+    subject: ['Accessible book', 'Protected DAISY'],
+    title: 'Nich \u00a3i \u0141a na karuseli',
+    ia_collection_s: 'printdisabled;china;internetarchivebooks',
+    printdisabled_s: 'OL26349201M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['Moskva'],
+    ia_box_id: ['IA1100601'],
+    edition_count: 1,
+    key: '/works/OL17751318W',
+    public_scan_b: false,
+    publisher: ['E ksmo'],
+    language: ['rus'],
+    last_modified_i: 1499402005,
+    first_publish_year: 2009,
+    cover_edition_key: 'OL26349201M',
+    publish_year: [2009],
+    publish_date: ['2009'],
+  },
+  {
+    title_suggest: 'After Dark',
+    edition_key: ['OL26047097M'],
+    cover_i: 7626177,
+    isbn: ['9788483831014', '8483831015'],
+    has_fulltext: true,
+    text: [
+      'OL26047097M',
+      '9788483831014',
+      '8483831015',
+      'Haruki Murakami',
+      'isbn_9788483831014',
+      'OL5036666A',
+      'Modelos',
+      'Students',
+      'Accessible book',
+      'Hermanas',
+      'Models (Persons)',
+      'Fiction',
+      'Estudiantes',
+      'Novela',
+      'Protected DAISY',
+      'Sisters',
+      'After Dark',
+      '/works/OL17461970W',
+      'TusQuets',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL26047097M',
+      '/works/OL17461970W',
+      '/subjects/modelos',
+      '/subjects/students',
+      '/subjects/hermanas',
+      '/subjects/models_(persons)',
+      '/subjects/fiction',
+      '/subjects/estudiantes',
+      '/subjects/novela',
+      '/subjects/sisters',
+      '/authors/OL5036666A',
+    ],
+    ia: ['isbn_9788483831014'],
+    author_key: ['OL5036666A'],
+    subject: [
+      'Modelos',
+      'Students',
+      'Accessible book',
+      'Hermanas',
+      'Models (Persons)',
+      'Fiction',
+      'Estudiantes',
+      'Novela',
+      'Protected DAISY',
+      'Sisters',
+    ],
+    title: 'After Dark',
+    lending_identifier_s: 'isbn_9788483831014',
+    ia_collection_s:
+      'printdisabled;browserlending;china;inlibrary;internetarchivebooks',
+    printdisabled_s: 'OL26047097M',
+    type: 'work',
+    ebook_count_i: 1,
+    ia_box_id: ['IA1139402'],
+    edition_count: 1,
+    first_publish_year: 2008,
+    key: '/works/OL17461970W',
+    public_scan_b: false,
+    publisher: ['TusQuets'],
+    last_modified_i: 1498322070,
+    lending_edition_s: 'OL26047097M',
+    cover_edition_key: 'OL26047097M',
+    publish_year: [2008],
+    publish_date: ['2008'],
+  },
+  {
+    title_suggest: 'To kyo  kitanshu',
+    edition_key: ['OL26342658M'],
+    cover_i: 8024598,
+    isbn: ['4101001561', '9784101001562'],
+    has_fulltext: true,
+    text: [
+      'OL26342658M',
+      '4101001561',
+      '9784101001562',
+      'Haruki Murakami',
+      '187709085',
+      'tokyokitancollec00haru',
+      'OL7087900A',
+      'Social life and customs',
+      'Accessible book',
+      'Protected DAISY',
+      'City and town life',
+      'Fiction',
+      'To kyo  kitanshu',
+      '/works/OL17744154W',
+      'Murakami Haruki cho',
+      'Shincho sha',
+      'Japan',
+      'Tokyo (Japan)',
+      'Tokyo',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL26342658M',
+      '/works/OL17744154W',
+      '/subjects/social_life_and_customs',
+      '/subjects/city_and_town_life',
+      '/subjects/fiction',
+      '/subjects/place:japan',
+      '/subjects/place:tokyo_(japan)',
+      '/subjects/place:tokyo',
+      '/authors/OL7087900A',
+    ],
+    oclc: ['187709085'],
+    ia: ['tokyokitancollec00haru'],
+    author_key: ['OL7087900A'],
+    subject: [
+      'Social life and customs',
+      'Accessible book',
+      'Protected DAISY',
+      'City and town life',
+      'Fiction',
+    ],
+    title: 'To kyo  kitanshu',
+    lending_identifier_s: 'tokyokitancollec00haru',
+    ia_collection_s:
+      'printdisabled;browserlending;china;inlibrary;internetarchivebooks',
+    printdisabled_s: 'OL26342658M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['To kyo'],
+    ia_box_id: ['IA153701'],
+    edition_count: 1,
+    first_publish_year: 2007,
+    key: '/works/OL17744154W',
+    public_scan_b: false,
+    publisher: ['Shincho sha'],
+    language: ['jpn'],
+    last_modified_i: 1498335306,
+    lending_edition_s: 'OL26342658M',
+    cover_edition_key: 'OL26342658M',
+    publish_year: [2007],
+    publish_date: ['2007'],
+    place: ['Japan', 'Tokyo (Japan)', 'Tokyo'],
+  },
+  {
+    title_suggest: 'Uzumaki-neko no mitsukekata',
+    edition_key: ['OL26341490M'],
+    cover_i: 8023024,
+    isbn: ['9784103534068', '4103534060'],
+    has_fulltext: true,
+    text: [
+      'OL26341490M',
+      '9784103534068',
+      '4103534060',
+      'Haruki Murakami',
+      'Murakami, Yo ko',
+      'Anzai, Mizumaru, 1942-2014',
+      '35659705',
+      'isbn_9784103534068',
+      'OL7087900A',
+      'Accessible book',
+      'Protected DAISY',
+      'Murakami Asahido  ja naru',
+      'Uzumaki-neko no mitsukekata',
+      '/works/OL17742788W',
+      'Murakami Haruki ; e, Anzai Mizumaru ; shashin, Murakami Yo ko',
+      'Shincho sha',
+    ],
+    author_name: ['Haruki Murakami'],
+    contributor: ['Murakami, Yo ko', 'Anzai, Mizumaru, 1942-2014'],
+    seed: ['/books/OL26341490M', '/works/OL17742788W', '/authors/OL7087900A'],
+    oclc: ['35659705'],
+    ia: ['isbn_9784103534068'],
+    author_key: ['OL7087900A'],
+    subject: ['Accessible book', 'Protected DAISY'],
+    title: 'Uzumaki-neko no mitsukekata',
+    lending_identifier_s: 'isbn_9784103534068',
+    ia_collection_s:
+      'printdisabled;browserlending;china;inlibrary;internetarchivebooks',
+    printdisabled_s: 'OL26341490M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['To kyo'],
+    ia_box_id: ['IA1113324'],
+    edition_count: 1,
+    first_publish_year: 1996,
+    key: '/works/OL17742788W',
+    public_scan_b: false,
+    publisher: ['Shincho sha'],
+    language: ['jpn'],
+    last_modified_i: 1498328842,
+    lending_edition_s: 'OL26341490M',
+    cover_edition_key: 'OL26341490M',
+    publish_year: [1996],
+    publish_date: ['1996'],
+  },
+  {
+    title_suggest: 'Sanshiro',
+    publisher: ['Penguin Books'],
+    cover_i: 7601828,
+    isbn: ['9780140455625', '0140455620'],
+    has_fulltext: false,
+    title: 'Sanshiro',
+    edition_key: ['OL26034174M'],
+    last_modified_i: 1476435539,
+    edition_count: 1,
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26034174M',
+    seed: ['/books/OL26034174M', '/works/OL17449792W', '/authors/OL5036666A'],
+    first_publish_year: 2010,
+    publish_year: [2010],
+    key: '/works/OL17449792W',
+    text: [
+      'OL26034174M',
+      '9780140455625',
+      '0140455620',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Sanshiro',
+      '/works/OL17449792W',
+      'Penguin Books',
+    ],
+    publish_date: ['2010'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'Shichigatsu Kugatsu',
+    publisher: ['Shinchosha/Tsai Fong Books'],
+    cover_i: 7638841,
+    isbn: ['9784101001623', '4101001626'],
+    has_fulltext: false,
+    title: 'Shichigatsu Kugatsu',
+    edition_key: ['OL26053820M'],
+    last_modified_i: 1476442548,
+    edition_count: 1,
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26053820M',
+    seed: ['/books/OL26053820M', '/works/OL17468319W', '/authors/OL5036666A'],
+    first_publish_year: 2012,
+    publish_year: [2012],
+    key: '/works/OL17468319W',
+    text: [
+      'OL26053820M',
+      '9784101001623',
+      '4101001626',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Shichigatsu Kugatsu',
+      '/works/OL17468319W',
+      'Shinchosha/Tsai Fong Books',
+    ],
+    publish_date: ['2012'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'Kafka En La Orilla',
+    publisher: ['TusQuets'],
+    cover_i: 7710194,
+    isbn: ['9788483835241', '848383524X'],
+    has_fulltext: false,
+    title: 'Kafka En La Orilla',
+    edition_key: ['OL26092125M'],
+    last_modified_i: 1476457160,
+    edition_count: 1,
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26092125M',
+    seed: ['/books/OL26092125M', '/works/OL17504215W', '/authors/OL5036666A'],
+    first_publish_year: 2008,
+    publish_year: [2008],
+    key: '/works/OL17504215W',
+    text: [
+      'OL26092125M',
+      '9788483835241',
+      '848383524X',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Kafka En La Orilla',
+      '/works/OL17504215W',
+      'TusQuets',
+    ],
+    publish_date: ['2008'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'Norwegian Wood',
+    publisher: ['Vintage Books USA'],
+    cover_i: 7771569,
+    isbn: ['0307744663', '9780307744661'],
+    has_fulltext: false,
+    title: 'Norwegian Wood',
+    edition_key: ['OL26124589M'],
+    last_modified_i: 1476469856,
+    edition_count: 1,
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26124589M',
+    seed: ['/books/OL26124589M', '/works/OL17534964W', '/authors/OL5036666A'],
+    key: '/works/OL17534964W',
+    text: [
+      'OL26124589M',
+      '0307744663',
+      '9780307744661',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Norwegian Wood',
+      '/works/OL17534964W',
+      'Vintage Books USA',
+    ],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: '1Q84 Book 3',
+    publisher: ['Shi Bao Chu Ban/Tsai Fong Books'],
+    cover_i: 7774029,
+    isbn: ['9789571352497', '9571352497'],
+    has_fulltext: false,
+    title: '1Q84 Book 3',
+    edition_key: ['OL26125873M'],
+    last_modified_i: 1476470361,
+    edition_count: 1,
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26125873M',
+    seed: ['/books/OL26125873M', '/works/OL17536192W', '/authors/OL5036666A'],
+    first_publish_year: 2010,
+    publish_year: [2010],
+    key: '/works/OL17536192W',
+    text: [
+      'OL26125873M',
+      '9789571352497',
+      '9571352497',
+      'Haruki Murakami',
+      'OL5036666A',
+      '1Q84 Book 3',
+      '/works/OL17536192W',
+      'Shi Bao Chu Ban/Tsai Fong Books',
+    ],
+    publish_date: ['2010'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest:
+      'Al Sur de la Frontera al Oeste del Sol\r\n            \r\n                Coleccion Maxi',
+    publisher: ['TusQuets'],
+    cover_i: 7805974,
+    isbn: ['9788483835036', '8483835037'],
+    has_fulltext: false,
+    title:
+      'Al Sur de la Frontera al Oeste del Sol\r\n            \r\n                Coleccion Maxi',
+    edition_key: ['OL26142417M'],
+    last_modified_i: 1476477089,
+    edition_count: 1,
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26142417M',
+    seed: ['/books/OL26142417M', '/works/OL17552174W', '/authors/OL5036666A'],
+    first_publish_year: 2007,
+    publish_year: [2007],
+    key: '/works/OL17552174W',
+    text: [
+      'OL26142417M',
+      '9788483835036',
+      '8483835037',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Al Sur de la Frontera al Oeste del Sol\r\n            \r\n                Coleccion Maxi',
+      '/works/OL17552174W',
+      'TusQuets',
+    ],
+    publish_date: ['2007'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'Norwegian Wood',
+    publisher: ['Vintage Books'],
+    cover_i: 7822386,
+    isbn: ['030795062X', '9780307950628'],
+    has_fulltext: false,
+    title: 'Norwegian Wood',
+    edition_key: ['OL26150873M'],
+    last_modified_i: 1476481158,
+    edition_count: 1,
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26150873M',
+    seed: ['/books/OL26150873M', '/works/OL17560402W', '/authors/OL5036666A'],
+    first_publish_year: 2012,
+    publish_year: [2012],
+    key: '/works/OL17560402W',
+    text: [
+      'OL26150873M',
+      '030795062X',
+      '9780307950628',
+      'Haruki Murakami',
+      'OL5036666A',
+      'Norwegian Wood',
+      '/works/OL17560402W',
+      'Vintage Books',
+    ],
+    publish_date: ['2012'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'Hotaru, naya o yaku, sonota no tanpen',
+    edition_key: ['OL26197314M'],
+    cover_i: 7881092,
+    isbn: ['4101001332', '9784101001333'],
+    has_fulltext: true,
+    text: [
+      'OL26197314M',
+      '4101001332',
+      '9784101001333',
+      'Haruki Murakami',
+      '21920124',
+      'isbn_9784101001333',
+      'OL7087900A',
+      'Japanese fiction',
+      'In library',
+      'Accessible book',
+      'Protected DAISY',
+      'Hotaru, naya o yaku, sonota no tanpen',
+      '/works/OL17594043W',
+      'Murakami Haruki cho',
+      'Shincho sha',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL26197314M',
+      '/works/OL17594043W',
+      '/subjects/japanese_fiction',
+      '/subjects/in_library',
+      '/authors/OL7087900A',
+    ],
+    oclc: ['21920124'],
+    ia: ['isbn_9784101001333'],
+    author_key: ['OL7087900A'],
+    subject: [
+      'Japanese fiction',
+      'In library',
+      'Accessible book',
+      'Protected DAISY',
+    ],
+    title: 'Hotaru, naya o yaku, sonota no tanpen',
+    lending_identifier_s: 'isbn_9784101001333',
+    ia_collection_s:
+      'printdisabled;browserlending;china;inlibrary;internetarchivebooks',
+    printdisabled_s: 'OL26197314M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['To kyo'],
+    ia_box_id: ['IA1150404'],
+    edition_count: 1,
+    first_publish_year: 1987,
+    key: '/works/OL17594043W',
+    public_scan_b: false,
+    publisher: ['Shincho sha'],
+    language: ['jpn'],
+    last_modified_i: 1500704138,
+    lending_edition_s: 'OL26197314M',
+    cover_edition_key: 'OL26197314M',
+    publish_year: [1987],
+    publish_date: ['1987'],
+  },
+  {
+    title_suggest: 'De qu\u00e9 hablo cuando hablo de escribir',
+    publisher: ['Tusquets'],
+    isbn: ['8490663998', '978-84-9066-399-8'],
+    has_fulltext: false,
+    edition_count: 1,
+    edition_key: ['OL26246447M'],
+    last_modified_i: 1493107497,
+    title: 'De qu\u00e9 hablo cuando hablo de escribir',
+    author_name: ['Haruki Murakami'],
+    cover_edition_key: 'OL26246447M',
+    seed: ['/books/OL26246447M', '/works/OL17640709W', '/authors/OL5036666A'],
+    first_publish_year: 2017,
+    publish_year: [2017],
+    key: '/works/OL17640709W',
+    text: [
+      'OL26246447M',
+      '8490663998',
+      '978-84-9066-399-8',
+      'Haruki Murakami',
+      'OL5036666A',
+      'De qu\u00e9 hablo cuando hablo de escribir',
+      '/works/OL17640709W',
+      'Tusquets',
+    ],
+    publish_date: ['2017'],
+    author_key: ['OL5036666A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: 'Hashiru koto ni tsuite kataru toki ni boku no kataru koto',
+    edition_key: ['OL26388765M'],
+    cover_i: 8083014,
+    isbn: ['4167502100', '9784167502102'],
+    has_fulltext: true,
+    text: [
+      'OL26388765M',
+      '4167502100',
+      '9784167502102',
+      'Haruki Murakami',
+      '641460424',
+      'isbn_9784167502102',
+      'OL7087900A',
+      'Accessible book',
+      'Protected DAISY',
+      'Marathon running',
+      'Hashiru koto ni tsuite kataru toki ni boku no kataru koto',
+      '/works/OL17799221W',
+      'Murakami Haruki',
+      'Bungei Shunju',
+      'Haruki Murakami (1949-)',
+    ],
+    author_name: ['Haruki Murakami'],
+    seed: [
+      '/books/OL26388765M',
+      '/works/OL17799221W',
+      '/subjects/marathon_running',
+      '/subjects/person:haruki_murakami_(1949-)',
+      '/authors/OL7087900A',
+    ],
+    oclc: ['641460424'],
+    ia: ['isbn_9784167502102'],
+    author_key: ['OL7087900A'],
+    subject: ['Accessible book', 'Protected DAISY', 'Marathon running'],
+    title: 'Hashiru koto ni tsuite kataru toki ni boku no kataru koto',
+    ia_collection_s: 'printdisabled;china;internetarchivebooks',
+    printdisabled_s: 'OL26388765M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['To kyo'],
+    ia_box_id: ['IA1112901'],
+    edition_count: 1,
+    key: '/works/OL17799221W',
+    public_scan_b: false,
+    publisher: ['Bungei Shunju'],
+    language: ['jpn'],
+    last_modified_i: 1508390622,
+    first_publish_year: 2010,
+    cover_edition_key: 'OL26388765M',
+    person: ['Haruki Murakami (1949-)'],
+    publish_year: [2010],
+    publish_date: ['2010'],
+  },
+  {
+    title_suggest: 'Yoru no kumozaru',
+    edition_key: ['OL26424208M'],
+    cover_i: 8131918,
+    isbn: ['4101001448', '9784101001449'],
+    has_fulltext: true,
+    text: [
+      'OL26424208M',
+      '4101001448',
+      '9784101001449',
+      'Haruki Murakami',
+      'Anzai, Mizumaru, 1942-2014',
+      '45007748',
+      'yorunokumozarumu00008800',
+      'OL7087900A',
+      'Accessible book',
+      'Protected DAISY',
+      'Murakami Asahid\u014d ch\u014dtanpen sh\u014dsetsu',
+      'Yoru no kumozaru',
+      '/works/OL17837258W',
+      'Murakami Haruki, Anzai Mizumaru cho',
+      'Shinch\u014dsha',
+    ],
+    author_name: ['Haruki Murakami'],
+    contributor: ['Anzai, Mizumaru, 1942-2014'],
+    seed: ['/books/OL26424208M', '/works/OL17837258W', '/authors/OL7087900A'],
+    oclc: ['45007748'],
+    ia: ['yorunokumozarumu00008800'],
+    author_key: ['OL7087900A'],
+    subject: ['Accessible book', 'Protected DAISY'],
+    title: 'Yoru no kumozaru',
+    ia_collection_s: 'printdisabled;china;internetarchivebooks',
+    printdisabled_s: 'OL26424208M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL17837258W',
+    public_scan_b: false,
+    publisher: ['Shinch\u014dsha'],
+    language: ['jpn'],
+    last_modified_i: 1517892584,
+    first_publish_year: 1998,
+    cover_edition_key: 'OL26424208M',
+    publish_year: [1998],
+    publish_date: ['1998'],
+  },
+  {
+    title_suggest: 'Afut\u0101d\u0101ku =',
+    edition_key: ['OL21013494M', 'OL22997075M', 'OL23062452M'],
+    subtitle: 'Afterdark',
+    has_fulltext: false,
+    text: [
+      'OL21013494M',
+      'OL22997075M',
+      'OL23062452M',
+      'Afterdark',
+      'Murakami Haruki',
+      'Lin, Shaohua.',
+      'Lai, Mingzhu.',
+      '56611687',
+      '60528702',
+      '9789628871865',
+      '9571342475',
+      '9789571342474',
+      '9784062125369',
+      '9628871862',
+      '9787532736850',
+      '7532736857',
+      '4062125366',
+      'OL382524A',
+      'Afut\u0101d\u0101ku =',
+      'Cunshang Chunshu zhu ; Lin Shaohua yi.',
+      'Cunshang Chunshu zhu ; Lai Mingzhu yi.',
+      'Murakami Haruki.',
+      '/works/OL2625519W',
+      'Shanghai yi wen chu ban she',
+      'K\u014ddansha',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+      'Afut\u0101 d\u0101ku.',
+      'Afut\u0101 d\u0101ku',
+      'Hei ye zhi hou',
+      'Tian hei yi hou =',
+      'Afterdark',
+      'After dark',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL21013494M',
+      '/books/OL22997075M',
+      '/books/OL23062452M',
+      '/works/OL2625519W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: ['Lin, Shaohua.', 'Lai, Mingzhu.'],
+    isbn: [
+      '9789628871865',
+      '9571342475',
+      '9789571342474',
+      '9784062125369',
+      '9628871862',
+      '9787532736850',
+      '7532736857',
+      '4062125366',
+    ],
+    author_key: ['OL382524A'],
+    title: 'Afut\u0101d\u0101ku =',
+    publish_date: ['2004', '2005'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Shanghai', 'Taibei Shi'],
+    edition_count: 3,
+    key: '/works/OL2625519W',
+    id_goodreads: ['6304705', '1675168'],
+    publisher: [
+      'Shanghai yi wen chu ban she',
+      'K\u014ddansha',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+    ],
+    language: ['chi', 'jpn'],
+    last_modified_i: 1303977766,
+    id_librarything: ['637581'],
+    publish_year: [2004, 2005],
+    first_publish_year: 2004,
+    oclc: ['56611687', '60528702'],
+  },
+  {
+    title_suggest: 'Kami no kodomotachi wa mina odoru',
+    edition_key: ['OL19725114M', 'OL23056989M', 'OL23009072M'],
+    cover_i: 5490315,
+    subtitle:
+      'Murak\u02bbami Haruk\u02bbi ch\u02bboech\u02bbo \u016di y\u014fnjak sos\u014fl',
+    has_fulltext: false,
+    text: [
+      'OL19725114M',
+      'OL23056989M',
+      'OL23009072M',
+      'Murak\u02bbami Haruk\u02bbi ch\u02bboech\u02bbo \u016di y\u014fnjak sos\u014fl',
+      'Murakami Haruki',
+      'Lai, Mingzhu.',
+      '45665833',
+      '43797498',
+      '9784103534112',
+      '8970123598',
+      '4103534117',
+      '9571331759',
+      '9789571331751',
+      '9788970123592',
+      'OL382524A',
+      'Kami no kodomotachi wa mina odoru',
+      'Cunshang Chunshu zhu ; Lai Mingzhu yi.',
+      'Kim Yu-gon omgim.',
+      'Murakami Haruki.',
+      '/works/OL2625521W',
+      'Shinch\u014dsha',
+      'Munhak Sasangsa',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+      'Kami no kodomo tachi wa mina odoru.',
+      'Kami no kodomotachi wa mina odoru.',
+      'Shen de hai zi dou zai tiao wu',
+      'Sin \u016di aid\u016dl \u016dn modu ch\u02bbum ch\u02bbunda',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL19725114M',
+      '/books/OL23056989M',
+      '/books/OL23009072M',
+      '/works/OL2625521W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: ['Lai, Mingzhu.'],
+    isbn: [
+      '9784103534112',
+      '8970123598',
+      '4103534117',
+      '9571331759',
+      '9789571331751',
+      '9788970123592',
+    ],
+    author_key: ['OL382524A'],
+    title: 'Kami no kodomotachi wa mina odoru',
+    publish_date: ['2000'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: [
+      'S\u014ful T\u02bb\u016dkpy\u014fsi',
+      'T\u014dky\u014d',
+      'Taibei Shi',
+    ],
+    edition_count: 3,
+    key: '/works/OL2625521W',
+    id_goodreads: ['1373364'],
+    publisher: [
+      'Shinch\u014dsha',
+      'Munhak Sasangsa',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+    ],
+    language: ['chi', 'kor', 'jpn'],
+    last_modified_i: 1304113422,
+    id_librarything: ['31994'],
+    cover_edition_key: 'OL19725114M',
+    publish_year: [2000],
+    first_publish_year: 2000,
+    oclc: ['45665833', '43797498'],
+  },
+  {
+    title_suggest: 'Hitsujiotoko no Kurisumasu',
+    edition_key: ['OL22997946M', 'OL18508422M', 'OL23092325M'],
+    isbn: [
+      '9784062023634',
+      '9785699050543',
+      '569905054X',
+      '4061845764',
+      '9784061845763',
+      '4062023636',
+    ],
+    has_fulltext: false,
+    text: [
+      'OL22997946M',
+      'OL18508422M',
+      'OL23092325M',
+      '9784062023634',
+      '9785699050543',
+      '569905054X',
+      '4061845764',
+      '9784061845763',
+      '4062023636',
+      'Murakami Haruki',
+      'Sasaki, Maki.',
+      'Sasaki, Maki, 1946-',
+      'Zamilov, Andre\u012d.',
+      '26030387',
+      'OL382524A',
+      'Hitsujiotoko no Kurisumasu',
+      '[bun] Murakami Haruki ; [e] Sasaki Maki.',
+      'Kharuki Murakami ; illi\ufe20u\ufe21strat\ufe20s\ufe21ii Sasaki Maki ; perevod s i\ufe20a\ufe21ponskogo Andrei\ufe20a\ufe21 Zamilova.',
+      "Children's stories, Japanese",
+      'Translations into Russian',
+      '/works/OL2625520W',
+      'Izd-vo \u0116KSMO"',
+      'K\u014ddansha',
+      'Hitsujiotoko no Kurisumasu.',
+      'Rozhdestvo Ovt\ufe20s\ufe21y',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22997946M',
+      '/books/OL18508422M',
+      '/books/OL23092325M',
+      '/works/OL2625520W',
+      "/subjects/children's_stories_japanese",
+      '/subjects/translations_into_russian',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: [
+      'Sasaki, Maki.',
+      'Sasaki, Maki, 1946-',
+      'Zamilov, Andre\u012d.',
+    ],
+    author_key: ['OL382524A'],
+    subject: ["Children's stories, Japanese", 'Translations into Russian'],
+    title: 'Hitsujiotoko no Kurisumasu',
+    publish_date: ['1985', '2004', '1989'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Moskva'],
+    edition_count: 3,
+    key: '/works/OL2625520W',
+    id_goodreads: ['2376315', '2376313', '2858293'],
+    publisher: ['Izd-vo \u0116KSMO"', 'K\u014ddansha'],
+    language: ['rus', 'jpn'],
+    last_modified_i: 1282216551,
+    id_librarything: ['3011610'],
+    publish_year: [1985, 2004, 1989],
+    first_publish_year: 1985,
+    oclc: ['26030387'],
+  },
+  {
+    title_suggest: 'Henky\u014d, kinky\u014d',
+    edition_key: ['OL22971986M', 'OL23008007M', 'OL19365049M'],
+    subtitle: 'Ts\u02bbun-shang Ch\u02bbun-shu tso p\u02bbin chi',
+    has_fulltext: false,
+    text: [
+      'OL22971986M',
+      'OL23008007M',
+      'OL19365049M',
+      'Ts\u02bbun-shang Ch\u02bbun-shu tso p\u02bbin chi',
+      'Murakami Haruki',
+      '43304242',
+      '49255855',
+      '9784101001487',
+      '9789571328614',
+      '9784103534082',
+      '4103534087',
+      '4101001480',
+      '9571328618',
+      'OL382524A',
+      'Henky\u014d, kinky\u014d',
+      'Ts\u02bbun-shang Ch\u02bbun-shu chu ; Lai Ming-chu i.',
+      'Murakami Haruki.',
+      'Voyages and travels',
+      'Travel',
+      'Japanese Authors',
+      'Biography',
+      'Journeys',
+      '/works/OL2625507W',
+      'Shih pao wen hua ch\u02bbu pan ch\u02bbi yeh ku fen yu hsien kung ssu',
+      'Shinch\u014dsha',
+      'Bian jing, jin jing',
+      'Essays',
+      'Henky\u014d kinky\u014d.',
+      'Pien ching, chin ching',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Haruki Murakami (1949-)',
+      '20th century',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22971986M',
+      '/books/OL23008007M',
+      '/books/OL19365049M',
+      '/works/OL2625507W',
+      '/subjects/biography',
+      '/subjects/japanese_authors',
+      '/subjects/journeys',
+      '/subjects/travel',
+      '/subjects/voyages_and_travels',
+      '/subjects/person:haruki_murakami_(1949-)',
+      '/subjects/time:20th_century',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    isbn: [
+      '9784101001487',
+      '9789571328614',
+      '9784103534082',
+      '4103534087',
+      '4101001480',
+      '9571328618',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Voyages and travels',
+      'Travel',
+      'Japanese Authors',
+      'Biography',
+      'Journeys',
+    ],
+    title: 'Henky\u014d, kinky\u014d',
+    publish_date: ['1999', '1998', '2000'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'T\u02bbai-pei shih'],
+    edition_count: 3,
+    key: '/works/OL2625507W',
+    id_goodreads: ['1937567', '1854877'],
+    publisher: [
+      'Shih pao wen hua ch\u02bbu pan ch\u02bbi yeh ku fen yu hsien kung ssu',
+      'Shinch\u014dsha',
+    ],
+    language: ['chi', 'jpn'],
+    last_modified_i: 1304129782,
+    id_librarything: ['3089327', '9705701'],
+    cover_edition_key: 'OL19365049M',
+    person: ['Haruki Murakami (1949-)'],
+    publish_year: [1999, 1998, 2000],
+    first_publish_year: 1998,
+    time: ['20th century'],
+    oclc: ['43304242', '49255855'],
+  },
+  {
+    title_suggest: 'Kokky\u014d no minami, taiy\u014d no nishi',
+    edition_key: ['OL23053719M', 'OL22971821M', 'OL23074951M'],
+    isbn: [
+      '4062060817',
+      '9621711827',
+      '9784062060813',
+      '9789621711823',
+      '5699030506',
+      '9785699030507',
+    ],
+    has_fulltext: false,
+    text: [
+      'OL23053719M',
+      'OL22971821M',
+      'OL23074951M',
+      '4062060817',
+      '9621711827',
+      '9784062060813',
+      '9789621711823',
+      '5699030506',
+      '9785699030507',
+      'Murakami Haruki',
+      'Lai, Mingzhu.',
+      '27167197',
+      '53082757',
+      'OL382524A',
+      'Kokky\u014d no minami, taiy\u014d no nishi',
+      'Kharuki Murakami ; [perevod s i\ufe20a\ufe21ponskogo Ivana i Sergei\ufe20a\ufe21 Logachevykh].',
+      'Ts\u02bbun-shang Ch\u02bbun-shu chu ; Lai Ming-chu i =  South of the border, west of the sun / by Haruki Murakami.',
+      'Murakami Haruki.',
+      '/works/OL2625523W',
+      '\u0116KSMO',
+      'K\u014ddansha',
+      'Po i ch\u02bbu pan chi t\u02bbuan yu hsien kung ssu',
+      'Kokky\u014d no minami, taiy\u014d no nishi.',
+      'Kokky\u014d no minami, Taiy\u014d no nishi.',
+      'Guo jing zhi nan, tai yang zhi xi',
+      'Kuo ching chih nan, t\u02bbai yang chih hsi',
+      'K I\ufe20U\ufe21gu ot granit\ufe20s\ufe21y, na zapad ot solnt\ufe20s\ufe21a',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL23053719M',
+      '/books/OL22971821M',
+      '/books/OL23074951M',
+      '/works/OL2625523W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: ['Lai, Mingzhu.'],
+    author_key: ['OL382524A'],
+    title: 'Kokky\u014d no minami, taiy\u014d no nishi',
+    publish_date: ['2003', '1993', '1992'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Moskva', 'Hsiang-kang'],
+    edition_count: 3,
+    key: '/works/OL2625523W',
+    id_goodreads: ['1712021', '3799604'],
+    publisher: [
+      '\u0116KSMO',
+      'K\u014ddansha',
+      'Po i ch\u02bbu pan chi t\u02bbuan yu hsien kung ssu',
+    ],
+    language: ['chi', 'rus', 'jpn'],
+    last_modified_i: 1282215606,
+    id_librarything: ['5042'],
+    publish_year: [2003, 1993, 1992],
+    first_publish_year: 1992,
+    oclc: ['27167197', '53082757'],
+  },
+  {
+    title_suggest: 'South of the border, west of the sun',
+    edition_key: ['OL701136M', 'OL7700911M', 'OL24240969M'],
+    cover_i: 225553,
+    subtitle: 'A Novel',
+    has_fulltext: true,
+    text: [
+      'OL701136M',
+      'OL7700911M',
+      'OL24240969M',
+      'A Novel',
+      'Murakami Haruki',
+      'Philip Gabriel (Translator)',
+      'southofborderwes00mura',
+      '0679767398',
+      '0099448572',
+      '9780099448570',
+      '9780679767398',
+      '9780375402517',
+      '0375402519',
+      'OL382524A',
+      'Accessible book',
+      'Protected DAISY',
+      'Fiction',
+      'South of the border, west of the sun',
+      '/works/OL2625514W',
+      'Haruki Murakami ; translated from the Japanese by Philip Gabriel.',
+      'Knopf',
+      'Distributed by Random House',
+      'Vintage',
+      'South of the Border, West of the Sun',
+      '97049459',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      "My birthday's the fourth of January, 1951.",
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Philip Gabriel (Translator)'],
+    seed: [
+      '/books/OL701136M',
+      '/books/OL7700911M',
+      '/books/OL24240969M',
+      '/works/OL2625514W',
+      '/subjects/fiction',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['5042'],
+    ia: ['southofborderwes00mura'],
+    isbn: [
+      '0679767398',
+      '0099448572',
+      '9780099448570',
+      '9780679767398',
+      '9780375402517',
+      '0375402519',
+    ],
+    author_key: ['OL382524A'],
+    subject: ['Accessible book', 'Protected DAISY', 'Fiction'],
+    title: 'South of the border, west of the sun',
+    ia_collection_s: 'printdisabled;china;internetarchivebooks',
+    printdisabled_s: 'OL7700911M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['New York', 'London, United Kingdom'],
+    ia_box_id: ['IA1150905'],
+    edition_count: 3,
+    key: '/works/OL2625514W',
+    id_goodreads: ['33263', '45311'],
+    public_scan_b: false,
+    publisher: ['Knopf', 'Distributed by Random House', 'Vintage'],
+    language: ['eng'],
+    lccn: ['97049459'],
+    last_modified_i: 1493993561,
+    first_publish_year: 1999,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL701136M',
+    first_sentence: ["My birthday's the fourth of January, 1951."],
+    publish_year: [1999, 2003, 2000],
+    publish_date: ['1999', 'March 14, 2000', '2003'],
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Yagate kanashiki gaikokugo',
+    edition_key: ['OL9029665M', 'OL17635696M', 'OL23119369M'],
+    cover_i: 7980060,
+    id_librarything: ['2163304'],
+    has_fulltext: true,
+    text: [
+      'OL9029665M',
+      'OL17635696M',
+      'OL23119369M',
+      '9784062634373',
+      '9784062068000',
+      '9571343943',
+      '4062068001',
+      '9789571343945',
+      '4062634376',
+      'Murakami Haruki',
+      'Lai, Mingzhu.',
+      '30699968',
+      '76260973',
+      '37348283',
+      'isbn_9784062068000',
+      'OL382524A',
+      'Description and travel',
+      'Journeys',
+      'Intellectual life',
+      'Accessible book',
+      'Japanese Authors',
+      'Authors, Japanese',
+      'Homes and haunts',
+      'Protected DAISY',
+      'Biography',
+      'Yagate kanashiki gaikokugo',
+      'Yagate kanashiki gaikokugo',
+      '/works/OL2625526W',
+      'Cunshang Chunshu zhu ; Lai Mingzhu yi.',
+      'Murakami Haruki.',
+      'Shi bao wen hua chu ban qi ye you xian gong si',
+      'K\u014ddansha',
+      'Kodansha',
+      'Zhong yu bei ai de wai guo yu =',
+      'Yagate kanashiki gaikokugo.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Haruki Murakami (1949-)',
+      'United States',
+      '20th century',
+    ],
+    author_name: ['Murakami Haruki'],
+    contributor: ['Lai, Mingzhu.'],
+    seed: [
+      '/books/OL9029665M',
+      '/books/OL17635696M',
+      '/books/OL23119369M',
+      '/works/OL2625526W',
+      '/subjects/authors_japanese',
+      '/subjects/biography',
+      '/subjects/homes_and_haunts',
+      '/subjects/intellectual_life',
+      '/subjects/japanese_authors',
+      '/subjects/description_and_travel',
+      '/subjects/journeys',
+      '/subjects/person:haruki_murakami_(1949-)',
+      '/subjects/place:united_states',
+      '/subjects/time:20th_century',
+      '/authors/OL382524A',
+    ],
+    oclc: ['30699968', '76260973', '37348283'],
+    ia: ['isbn_9784062068000'],
+    isbn: [
+      '9784062634373',
+      '9784062068000',
+      '9571343943',
+      '4062068001',
+      '9789571343945',
+      '4062634376',
+    ],
+    author_key: ['OL382524A'],
+    subject: [
+      'Description and travel',
+      'Journeys',
+      'Intellectual life',
+      'Accessible book',
+      'Japanese Authors',
+      'Authors, Japanese',
+      'Homes and haunts',
+      'Protected DAISY',
+      'Biography',
+    ],
+    title: 'Yagate kanashiki gaikokugo',
+    lending_identifier_s: 'isbn_9784062068000',
+    ia_collection_s:
+      'printdisabled;browserlending;china;inlibrary;internetarchivebooks',
+    printdisabled_s: 'OL9029665M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['T\u014dky\u014d', 'Taibei Shi'],
+    ia_box_id: ['IA1150515'],
+    edition_count: 3,
+    first_publish_year: 1994,
+    key: '/works/OL2625526W',
+    id_goodreads: ['3398965', '533275', '1711906'],
+    public_scan_b: false,
+    publisher: [
+      'Shi bao wen hua chu ban qi ye you xian gong si',
+      'K\u014ddansha',
+      'Kodansha',
+    ],
+    language: ['chi', 'jpn'],
+    last_modified_i: 1494017327,
+    lending_edition_s: 'OL9029665M',
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL9029665M',
+    person: ['Haruki Murakami (1949-)'],
+    publish_year: [2006, 1994, 1997],
+    publish_date: ['2006', '1994', '1997'],
+    place: ['United States'],
+    time: ['20th century'],
+  },
+  {
+    title_suggest: 'Kangar\u016b-biyori',
+    edition_key: ['OL2925992M', 'OL23103763M'],
+    isbn: ['9785699164264', '569916426X'],
+    has_fulltext: false,
+    text: [
+      'OL2925992M',
+      'OL23103763M',
+      '9785699164264',
+      '569916426X',
+      'Murakami Haruki',
+      '70125129',
+      'OL382524A',
+      'Kangar\u016b-biyori',
+      'Kharuki Murakami ; [perevod s i\ufe20a\ufe21ponskogo Sergei\ufe20a\ufe21 Logacheva].',
+      'Murakami Haruki.',
+      'Translations into Russian',
+      'Japanese Short stories',
+      'Short stories, Japanese',
+      '/works/OL2625508W',
+      'Heibonsha',
+      '\u0116KSMO',
+      'Khoroshi\u012d den\u02b9 dli\ufe20a\ufe21 kenguru',
+      'Kangar\u016b-biyori.',
+      '84160534',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL2925992M',
+      '/books/OL23103763M',
+      '/works/OL2625508W',
+      '/subjects/japanese_short_stories',
+      '/subjects/short_stories_japanese',
+      '/subjects/translations_into_russian',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    subject: [
+      'Translations into Russian',
+      'Japanese Short stories',
+      'Short stories, Japanese',
+    ],
+    title: 'Kangar\u016b-biyori',
+    publish_date: ['2006', '1983'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Moskva'],
+    edition_count: 2,
+    key: '/works/OL2625508W',
+    id_goodreads: ['151515'],
+    publisher: ['Heibonsha', '\u0116KSMO'],
+    language: ['rus', 'jpn'],
+    lccn: ['84160534'],
+    last_modified_i: 1291449842,
+    id_librarything: ['1249774', '8280878'],
+    publish_year: [2006, 1983],
+    first_publish_year: 1983,
+    oclc: ['70125129'],
+  },
+  {
+    title_suggest: 'And\u0101guraundo =',
+    edition_key: ['OL18539923M', 'OL23077976M'],
+    subtitle: 'Underground',
+    has_fulltext: false,
+    text: [
+      'OL18539923M',
+      'OL23077976M',
+      'Underground',
+      'Murakami Haruki',
+      '36708043',
+      '65520085',
+      '5699157700',
+      '9785699157709',
+      '4062085755',
+      '9784062085755',
+      'OL382524A',
+      'And\u0101guraundo =',
+      'Kharuki Murakami ; [perevod s i\ufe20a\ufe21ponskogo Andrei\ufe20a\ufe21 Zamilova i Feliksa Tymakhovicha].',
+      'Murakami Haruki.',
+      'Terrorism',
+      'Public opinion',
+      'Oumu Shinriky\u014d (Religious organization)',
+      'Crime',
+      '/works/OL2625501W',
+      'K\u014ddansha',
+      '\u0116ksmo',
+      'Underground',
+      'Podzemka',
+      'And\u0101guraundo.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL18539923M',
+      '/books/OL23077976M',
+      '/works/OL2625501W',
+      '/subjects/oumu_shinriky\u014d_(religious_organization)',
+      '/subjects/terrorism',
+      '/subjects/public_opinion',
+      '/subjects/crime',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    isbn: ['5699157700', '9785699157709', '4062085755', '9784062085755'],
+    author_key: ['OL382524A'],
+    subject: [
+      'Terrorism',
+      'Public opinion',
+      'Oumu Shinriky\u014d (Religious organization)',
+      'Crime',
+    ],
+    title: 'And\u0101guraundo =',
+    publish_date: ['2006', '1997'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Moskva'],
+    edition_count: 2,
+    key: '/works/OL2625501W',
+    id_goodreads: ['146928', '1937582'],
+    publisher: ['K\u014ddansha', '\u0116ksmo'],
+    language: ['rus', 'jpn'],
+    last_modified_i: 1303977648,
+    id_librarything: ['58240'],
+    publish_year: [2006, 1997],
+    first_publish_year: 1997,
+    place: ['Japan'],
+    oclc: ['36708043', '65520085'],
+  },
+  {
+    title_suggest: 'Sup\u016btoniku no koibito =',
+    edition_key: ['OL22977382M', 'OL23081182M'],
+    subtitle: 'The Sputnik sweetheart',
+    has_fulltext: false,
+    text: [
+      'OL22977382M',
+      'OL23081182M',
+      'The Sputnik sweetheart',
+      'Murakami Haruki',
+      '41333036',
+      '9650713530',
+      '9789650713539',
+      '4062096579',
+      '9784062096577',
+      'OL382524A',
+      'Sup\u016btoniku no koibito =',
+      'Haru\u1e33i Mura\u1e33ami ; me-Yapanit, \u02bbEnat \u1e32uper.',
+      'Murakami Haruki.',
+      'Maestros',
+      'Unrequited love',
+      'Novela',
+      'Missing persons',
+      'Fiction',
+      'Women novelists',
+      'Businesswomen',
+      'Teachers',
+      '/works/OL2625515W',
+      'Keter',
+      'K\u014ddansha',
+      'Sputnik sweetheart.',
+      'Spu\u1e6dni\u1e33 ahuvati',
+      'Sputnik sweetheart',
+      'Sup\u016btoniku no koibito.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22977382M',
+      '/books/OL23081182M',
+      '/works/OL2625515W',
+      '/subjects/fiction',
+      '/subjects/novela',
+      '/subjects/maestros',
+      '/subjects/missing_persons',
+      '/subjects/unrequited_love',
+      '/subjects/women_novelists',
+      '/subjects/businesswomen',
+      '/subjects/teachers',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    isbn: ['9650713530', '9789650713539', '4062096579', '9784062096577'],
+    author_key: ['OL382524A'],
+    subject: [
+      'Maestros',
+      'Unrequited love',
+      'Novela',
+      'Missing persons',
+      'Fiction',
+      'Women novelists',
+      'Businesswomen',
+      'Teachers',
+    ],
+    title: 'Sup\u016btoniku no koibito =',
+    publish_date: ['1999', '2005'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Yerushalayim'],
+    edition_count: 2,
+    key: '/works/OL2625515W',
+    id_goodreads: ['1712077'],
+    publisher: ['Keter', 'K\u014ddansha'],
+    language: ['heb', 'jpn'],
+    last_modified_i: 1320451571,
+    id_librarything: ['74326'],
+    publish_year: [1999, 2005],
+    first_publish_year: 1999,
+    place: ['Japan'],
+    oclc: ['41333036'],
+  },
+  {
+    title_suggest: 'Mister Aufziehvogel',
+    edition_key: ['OL9852589M', 'OL9035949M'],
+    cover_i: 1007413,
+    isbn: ['9783442726684', '3770144791', '9783770144792', '3442726689'],
+    has_fulltext: false,
+    text: [
+      'OL9852589M',
+      'OL9035949M',
+      '9783442726684',
+      '3770144791',
+      '9783770144792',
+      '3442726689',
+      'Murakami Haruki',
+      '75918771',
+      'OL382524A',
+      'Mister Aufziehvogel',
+      '/works/OL2625509W',
+      'DUMONT Literatur und Kunst Verlag',
+      'btb',
+      'Mister Aufziehvogel.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL9852589M',
+      '/books/OL9035949M',
+      '/works/OL2625509W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Mister Aufziehvogel',
+    publish_date: ['September 1, 1998', 'November 1, 2000'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 2,
+    key: '/works/OL2625509W',
+    id_goodreads: ['1477355', '45320'],
+    publisher: ['DUMONT Literatur und Kunst Verlag', 'btb'],
+    language: ['ger'],
+    last_modified_i: 1303510168,
+    id_librarything: ['5069'],
+    cover_edition_key: 'OL9035949M',
+    publish_year: [1998, 2000],
+    first_publish_year: 1998,
+    oclc: ['75918771'],
+  },
+  {
+    title_suggest: 'Gef\u00e4hrliche Geliebte',
+    edition_key: ['OL9091990M', 'OL9035979M'],
+    cover_i: 1007443,
+    isbn: ['3770147812', '3442727952', '9783770147816', '9783442727957'],
+    has_fulltext: false,
+    text: [
+      'OL9091990M',
+      'OL9035979M',
+      '3770147812',
+      '3442727952',
+      '9783770147816',
+      '9783442727957',
+      'Murakami Haruki',
+      'OL382524A',
+      'Gef\u00e4hrliche Geliebte',
+      '/works/OL2625418W',
+      'DUMONT Literatur und Kunst Verlag',
+      'btb',
+      'Gef\u00e4hrliche Geliebte.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL9091990M',
+      '/books/OL9035979M',
+      '/works/OL2625418W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Gef\u00e4hrliche Geliebte',
+    publish_date: ['February 1, 2000', 'July 1, 2002'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 2,
+    key: '/works/OL2625418W',
+    id_goodreads: ['1049215', '10143'],
+    publisher: ['DUMONT Literatur und Kunst Verlag', 'btb'],
+    language: ['ger'],
+    last_modified_i: 1282110082,
+    id_librarything: ['5042'],
+    cover_edition_key: 'OL9035979M',
+    publish_year: [2002, 2000],
+    first_publish_year: 2000,
+  },
+  {
+    title_suggest: "Au sud de la fronti\u00e8re, \u00e0 l'ouest du soleil",
+    edition_key: ['OL8932757M', 'OL8889562M'],
+    cover_i: 979092,
+    isbn: ['2714437486', '226403629X', '9782714437488', '9782264036292'],
+    has_fulltext: false,
+    text: [
+      'OL8932757M',
+      'OL8889562M',
+      '2714437486',
+      '226403629X',
+      '9782714437488',
+      '9782264036292',
+      'Murakami Haruki',
+      'OL382524A',
+      "Au sud de la fronti\u00e8re, \u00e0 l'ouest du soleil",
+      '/works/OL2625502W',
+      '10/18',
+      'Belfond',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL8932757M',
+      '/books/OL8889562M',
+      '/works/OL2625502W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: "Au sud de la fronti\u00e8re, \u00e0 l'ouest du soleil",
+    publish_date: ['February 7, 2002', 'February 6, 2003'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 2,
+    key: '/works/OL2625502W',
+    id_goodreads: ['146925', '1741517'],
+    publisher: ['10/18', 'Belfond'],
+    language: ['fre'],
+    last_modified_i: 1281444682,
+    id_librarything: ['5042'],
+    cover_edition_key: 'OL8889562M',
+    publish_year: [2002, 2003],
+    first_publish_year: 2002,
+  },
+  {
+    title_suggest: 'Panya saish\u016bgeki',
+    edition_key: ['OL22920482M', 'OL14790010M'],
+    isbn: ['9784163089300', '4163089306', '4167502011', '9784167502010'],
+    has_fulltext: false,
+    text: [
+      'OL22920482M',
+      'OL14790010M',
+      '9784163089300',
+      '4163089306',
+      '4167502011',
+      '9784167502010',
+      'Murakami Haruki',
+      '15332560',
+      'OL382524A',
+      'Panya saish\u016bgeki',
+      'Murakami Haruki.',
+      '/works/OL2625512W',
+      'Bungei Shunj\u016b',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22920482M',
+      '/books/OL14790010M',
+      '/works/OL2625512W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Panya saish\u016bgeki',
+    publish_date: ['1986', '1989'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['Tokyo'],
+    edition_count: 2,
+    key: '/works/OL2625512W',
+    id_goodreads: ['1854927', '1854924'],
+    publisher: ['Bungei Shunj\u016b'],
+    language: ['jpn'],
+    last_modified_i: 1304081385,
+    id_librarything: ['1674833'],
+    publish_year: [1986, 1989],
+    first_publish_year: 1986,
+    oclc: ['15332560'],
+  },
+  {
+    title_suggest: 'Fuwa fuwa',
+    edition_key: ['OL20812927M', 'OL23067668M'],
+    isbn: ['4062733358', '9784062733359'],
+    has_fulltext: false,
+    text: [
+      'OL20812927M',
+      'OL23067668M',
+      '4062733358',
+      '9784062733359',
+      'Murakami Haruki',
+      'Birnbaum, Alfred.',
+      'Sud\u014d, Kazuyoshi.',
+      'Anzai, Mizumaru, 1942-',
+      'OL382524A',
+      'Fuwa fuwa',
+      '[essay, Haruki Murakami ; texts, Kazuyoshi Sudo, Alfred Birnbaum].',
+      'Murakami Haruki bun ; Anzai Mizumaru e.',
+      'Textile design',
+      'Textile fabrics',
+      '/works/OL2625506W',
+      'Nuno Corporation',
+      'K\u014ddansha',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL20812927M',
+      '/books/OL23067668M',
+      '/works/OL2625506W',
+      '/subjects/textile_design',
+      '/subjects/textile_fabrics',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['2073300'],
+    contributor: [
+      'Birnbaum, Alfred.',
+      'Sud\u014d, Kazuyoshi.',
+      'Anzai, Mizumaru, 1942-',
+    ],
+    author_key: ['OL382524A'],
+    subject: ['Textile design', 'Textile fabrics'],
+    title: 'Fuwa fuwa',
+    publish_date: ['1998', '2001'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 2,
+    key: '/works/OL2625506W',
+    id_goodreads: ['1854884'],
+    publisher: ['Nuno Corporation', 'K\u014ddansha'],
+    language: ['jpn', 'eng'],
+    last_modified_i: 1282215290,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1998, 2001],
+    first_publish_year: 1998,
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Z\u014dk\u014dj\u014d no happ\u012b endo',
+    edition_key: ['OL22952254M', 'OL14789999M'],
+    isbn: ['4101001316', '9784101001319'],
+    has_fulltext: false,
+    text: [
+      'OL22952254M',
+      'OL14789999M',
+      '4101001316',
+      '9784101001319',
+      'Murakami Haruki',
+      'Anzai, Mizumaru, 1942-',
+      'OL382524A',
+      'Z\u014dk\u014dj\u014d no happ\u012b endo',
+      'Murakami Haruki, Anzai Mizumaru cho.',
+      'Murakami Haruki,  Anzai Mizumaru cho.',
+      '/works/OL2625518W',
+      'Shinch\u014dsha',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22952254M',
+      '/books/OL14789999M',
+      '/works/OL2625518W',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['9229181'],
+    contributor: ['Anzai, Mizumaru, 1942-'],
+    author_key: ['OL382524A'],
+    title: 'Z\u014dk\u014dj\u014d no happ\u012b endo',
+    publish_date: ['1986', '1983'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 2,
+    key: '/works/OL2625518W',
+    id_goodreads: ['1936981'],
+    publisher: ['Shinch\u014dsha'],
+    language: ['jpn'],
+    last_modified_i: 1282212204,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1986, 1983],
+    first_publish_year: 1983,
+  },
+  {
+    title_suggest: 'Murakami rajio',
+    edition_key: ['OL23037194M', 'OL23071060M'],
+    isbn: ['9789571337166', '9784838713141', '4838713142', '9571337161'],
+    has_fulltext: false,
+    text: [
+      'OL23037194M',
+      'OL23071060M',
+      '9789571337166',
+      '9784838713141',
+      '4838713142',
+      '9571337161',
+      'Murakami Haruki',
+      '\u014chashi, Ayumi.',
+      'Lai, Mingzhu.',
+      '47843065',
+      '50784927',
+      'OL382524A',
+      'Murakami rajio',
+      'Murakami Haruki bun ; \u014chashi Ayumi ga.',
+      'Cunshang Chunshu zhu ; Lai Mingzhu yi.',
+      '/works/OL2625510W',
+      'Magajin Hausu',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+      'Cunshang shou yin ji',
+      'Murakami rajio.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL23037194M',
+      '/books/OL23071060M',
+      '/works/OL2625510W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: ['\u014chashi, Ayumi.', 'Lai, Mingzhu.'],
+    author_key: ['OL382524A'],
+    title: 'Murakami rajio',
+    publish_date: ['2002', '2001'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Taibei Shi'],
+    edition_count: 2,
+    key: '/works/OL2625510W',
+    id_goodreads: ['1633710', '6902253'],
+    publisher: [
+      'Magajin Hausu',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+    ],
+    language: ['chi', 'jpn'],
+    last_modified_i: 1303822643,
+    id_librarything: ['7758459'],
+    publish_year: [2002, 2001],
+    first_publish_year: 2001,
+    oclc: ['47843065', '50784927'],
+  },
+  {
+    title_suggest: 'Ch\u016bgoku-yuki no sur\u014d b\u014dto',
+    edition_key: ['OL15953504M', 'OL23103716M'],
+    isbn: ['5699181245', '9785699181247'],
+    has_fulltext: false,
+    text: [
+      'OL15953504M',
+      'OL23103716M',
+      '5699181245',
+      '9785699181247',
+      'Murakami Haruki',
+      '71333936',
+      'OL382524A',
+      'Ch\u016bgoku-yuki no sur\u014d b\u014dto',
+      'Haruki Murakami ; [perevod s i\ufe20a\ufe21ponskogo Andrei\ufe20a\ufe21 Zamilova].',
+      'Murakami Haruki.',
+      'Short stories',
+      '/works/OL2625504W',
+      '\u0116KSMO',
+      'Ch\u016b\u014d K\u014dronsha',
+      'Ch\u016bgoku-yuki no sur\u014d b\u014dto.',
+      'Medlenno\u012d shli\ufe20u\ufe21pko\u012d v Kita\u012d',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL15953504M',
+      '/books/OL23103716M',
+      '/works/OL2625504W',
+      '/subjects/short_stories',
+      '/authors/OL382524A',
+    ],
+    oclc: ['71333936'],
+    author_key: ['OL382524A'],
+    subject: ['Short stories'],
+    title: 'Ch\u016bgoku-yuki no sur\u014d b\u014dto',
+    publish_date: ['2006', '1983'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Moskva'],
+    edition_count: 2,
+    key: '/works/OL2625504W',
+    id_goodreads: ['2524361'],
+    publisher: ['\u0116KSMO', 'Ch\u016b\u014d K\u014dronsha'],
+    language: ['rus', 'jpn'],
+    last_modified_i: 1282110079,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [2006, 1983],
+    first_publish_year: 1983,
+  },
+  {
+    title_suggest: 'Naokos L\u00e4cheln. Nur eine Liebesgeschichte',
+    edition_key: ['OL9440880M', 'OL9036013M'],
+    cover_i: 2237620,
+    isbn: ['9783442730506', '9783770156092', '3442730503', '3770156099'],
+    has_fulltext: false,
+    text: [
+      'OL9440880M',
+      'OL9036013M',
+      '9783442730506',
+      '9783770156092',
+      '3442730503',
+      '3770156099',
+      'Murakami Haruki',
+      '76281786',
+      'OL382524A',
+      'Naokos L\u00e4cheln. Nur eine Liebesgeschichte',
+      '/works/OL2625511W',
+      'Btb Bei Goldmann',
+      'DUMONT Literatur und Kunst Verlag',
+      'Naokos L\u00e4cheln. Nur eine Liebesgeschichte.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL9440880M',
+      '/books/OL9036013M',
+      '/works/OL2625511W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Naokos L\u00e4cheln. Nur eine Liebesgeschichte',
+    publish_date: ['February 1, 2003', 'January 1, 2001'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 2,
+    key: '/works/OL2625511W',
+    id_goodreads: ['45319', '9987'],
+    publisher: ['Btb Bei Goldmann', 'DUMONT Literatur und Kunst Verlag'],
+    language: ['ger'],
+    last_modified_i: 1304123971,
+    id_librarything: ['5860'],
+    cover_edition_key: 'OL9440880M',
+    publish_year: [2003, 2001],
+    first_publish_year: 2001,
+    oclc: ['76281786'],
+  },
+  {
+    title_suggest: 'Rekishinton no y\u016brei',
+    edition_key: ['OL22956139M', 'OL23004582M'],
+    subtitle: 'rasskazy',
+    has_fulltext: false,
+    text: [
+      'OL22956139M',
+      'OL23004582M',
+      'rasskazy',
+      'Murakami Haruki',
+      '36324755',
+      '52861747',
+      '9784163166308',
+      '4163166300',
+      '5699019375',
+      '9785699019373',
+      'OL382524A',
+      'Rekishinton no y\u016brei',
+      'Kharuki Murakami ; [pervod s i\ufe20a\ufe21ponskogo A. Zamilova].',
+      'Murakami Haruki.',
+      'Translations into Russian',
+      'Japanese Short stories',
+      '/works/OL2625513W',
+      'EKSMO',
+      'Bungei Shunj\u016b',
+      'Rekishinton no y\u016brei.',
+      'Prizraki Leksingtona',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22956139M',
+      '/books/OL23004582M',
+      '/works/OL2625513W',
+      '/subjects/translations_into_russian',
+      '/subjects/japanese_short_stories',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    isbn: ['9784163166308', '4163166300', '5699019375', '9785699019373'],
+    author_key: ['OL382524A'],
+    subject: ['Translations into Russian', 'Japanese Short stories'],
+    title: 'Rekishinton no y\u016brei',
+    publish_date: ['2003', '1996'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Moskva'],
+    edition_count: 2,
+    key: '/works/OL2625513W',
+    id_goodreads: ['1936727'],
+    publisher: ['EKSMO', 'Bungei Shunj\u016b'],
+    language: ['rus', 'jpn'],
+    last_modified_i: 1282212204,
+    id_librarything: ['9598624'],
+    publish_year: [2003, 1996],
+    first_publish_year: 1996,
+    oclc: ['36324755', '52861747'],
+  },
+  {
+    title_suggest: 'Fushigi na toshokan',
+    edition_key: ['OL23040384M', 'OL21133277M'],
+    isbn: ['4062127393', '9784062127394'],
+    has_fulltext: false,
+    text: [
+      'OL23040384M',
+      'OL21133277M',
+      '4062127393',
+      '9784062127394',
+      'Murakami Haruki',
+      'Sasaki, Maki, 1946-',
+      '60249977',
+      'OL382524A',
+      'Fushigi na toshokan',
+      '[bun] Murakami Haruki ; [e] Sasaki Maki.',
+      'Murakami Haruki ; Sasaki Maki.',
+      '/works/OL2625505W',
+      'K\u014ddansha',
+      'Fushigina toshokan',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL23040384M',
+      '/books/OL21133277M',
+      '/works/OL2625505W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: ['Sasaki, Maki, 1946-'],
+    author_key: ['OL382524A'],
+    title: 'Fushigi na toshokan',
+    publish_date: ['2005'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 2,
+    key: '/works/OL2625505W',
+    id_goodreads: ['2376347'],
+    publisher: ['K\u014ddansha'],
+    language: ['jpn'],
+    last_modified_i: 1304064409,
+    id_librarything: ['242943'],
+    publish_year: [2005],
+    first_publish_year: 2005,
+    oclc: ['60249977'],
+  },
+  {
+    title_suggest: 'Nuowei de sen lin',
+    edition_key: ['OL22960135M', 'OL23022274M'],
+    isbn: [
+      '9789621708960',
+      '9789571340067',
+      '9621708966',
+      '9789571340043',
+      '9621708958',
+      '9571340049',
+      '9789571340050',
+      '9789621708953',
+      '9571340065',
+      '9571340057',
+    ],
+    has_fulltext: false,
+    text: [
+      'OL22960135M',
+      'OL23022274M',
+      '9789621708960',
+      '9789571340067',
+      '9621708966',
+      '9789571340043',
+      '9621708958',
+      '9571340049',
+      '9789571340050',
+      '9789621708953',
+      '9571340065',
+      '9571340057',
+      'Murakami Haruki',
+      'Lai, Mingzhu.',
+      '35518089',
+      'OL382524A',
+      'Nuowei de sen lin',
+      'Cunshang Chunshu zhu ; Lai Mingzhu yi.',
+      'Cunshang Chunshu [zhu] ; Ye Hui yi.',
+      '/works/OL2625459W',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+      'Bo yi chu ban ji tuan you xian gong si',
+      'Noruwei no mori.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22960135M',
+      '/books/OL23022274M',
+      '/works/OL2625459W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: ['Lai, Mingzhu.'],
+    author_key: ['OL382524A'],
+    title: 'Nuowei de sen lin',
+    publish_date: ['1991', '2003'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['Taibei Shi', 'Xianggang'],
+    edition_count: 2,
+    key: '/works/OL2625459W',
+    id_goodreads: ['231031'],
+    publisher: [
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+      'Bo yi chu ban ji tuan you xian gong si',
+    ],
+    language: ['chi'],
+    last_modified_i: 1282213809,
+    id_librarything: ['5860'],
+    publish_year: [1991, 2003],
+    first_publish_year: 1991,
+    oclc: ['35518089'],
+  },
+  {
+    title_suggest: 'Yakusokusareta basho de =',
+    edition_key: ['OL22953905M', 'OL23092825M'],
+    subtitle: 'Underground 2',
+    has_fulltext: false,
+    text: [
+      'OL22953905M',
+      'OL23092825M',
+      'Underground 2',
+      'Murakami Haruki',
+      '40774404',
+      '5699199314',
+      '4163546006',
+      '9784163546001',
+      '9785699199310',
+      'OL382524A',
+      'Podzemka-2',
+      'Yakusokusareta basho de =',
+      'Kharuki Murakami ; [perevod s i\ufe20a\ufe21ponskogo S. Logacheva i A. Zamilova].',
+      'Murakami Haruki.',
+      'Cults',
+      'Terrorism',
+      'Oumu Shinriky\u014d (Religious organization)',
+      '/works/OL2625527W',
+      'Bungei Shunj\u016b',
+      '\u0116ksmo',
+      'Kra\u012d obetovanny\u012d',
+      'Podzemka-2',
+      'Yakusoku sareta basho de.',
+      'Underground 2',
+      'Podzemka-dva',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Japan',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22953905M',
+      '/books/OL23092825M',
+      '/works/OL2625527W',
+      '/subjects/oumu_shinriky\u014d_(religious_organization)',
+      '/subjects/cults',
+      '/subjects/terrorism',
+      '/subjects/place:japan',
+      '/authors/OL382524A',
+    ],
+    oclc: ['40774404'],
+    isbn: ['5699199314', '4163546006', '9784163546001', '9785699199310'],
+    author_key: ['OL382524A'],
+    subject: [
+      'Cults',
+      'Terrorism',
+      'Oumu Shinriky\u014d (Religious organization)',
+    ],
+    title: 'Yakusokusareta basho de =',
+    publish_date: ['1998', '2006'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d', 'Moskva'],
+    edition_count: 2,
+    key: '/works/OL2625527W',
+    id_goodreads: ['1819234', '2377076'],
+    publisher: ['Bungei Shunj\u016b', '\u0116ksmo'],
+    language: ['rus', 'jpn'],
+    last_modified_i: 1303513355,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1998, 2006],
+    first_publish_year: 1998,
+    place: ['Japan'],
+  },
+  {
+    title_suggest: 'Xun yang de mao xian',
+    publish_place: ['Hong Kong'],
+    edition_key: ['OL21180275M', 'OL21007141M'],
+    isbn: ['9621710723', '9621710731', '9789621710727', '9789621710734'],
+    has_fulltext: false,
+    language: ['chi'],
+    edition_count: 2,
+    oclc: ['27353734'],
+    last_modified_i: 1303932181,
+    title: 'Xun yang de mao xian',
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    seed: [
+      '/books/OL21180275M',
+      '/books/OL21007141M',
+      '/works/OL11519617W',
+      '/authors/OL382524A',
+    ],
+    author_name: ['Murakami Haruki'],
+    key: '/works/OL11519617W',
+    text: [
+      'OL21180275M',
+      'OL21007141M',
+      '9621710723',
+      '9621710731',
+      '9789621710727',
+      '9789621710734',
+      'Murakami Haruki',
+      '27353734',
+      'OL382524A',
+      'Xun yang de mao xian',
+      '/works/OL11519617W',
+      'Xun yang de mao xian .',
+      'Xun yang de mao xian.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_key: ['OL382524A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest: "La Ballade de l'impossible",
+    edition_key: ['OL8834305M', 'OL8835298M'],
+    cover_i: 963607,
+    isbn: ['2020579391', '9782020134736', '9782020579391', '202013473X'],
+    has_fulltext: false,
+    text: [
+      'OL8834305M',
+      'OL8835298M',
+      '2020579391',
+      '9782020134736',
+      '9782020579391',
+      '202013473X',
+      'Murakami Haruki',
+      '31261845',
+      'OL382524A',
+      "La Ballade de l'impossible",
+      '/works/OL15029004W',
+      'Seuil',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL8834305M',
+      '/books/OL8835298M',
+      '/works/OL15029004W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: "La Ballade de l'impossible",
+    publish_date: ['February 14, 2003', 'March 1, 1999'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 2,
+    key: '/works/OL15029004W',
+    id_goodreads: ['1049224', '1389017'],
+    publisher: ['Seuil'],
+    language: ['fre'],
+    last_modified_i: 1303960274,
+    id_librarything: ['5860'],
+    cover_edition_key: 'OL8834305M',
+    publish_year: [1999, 2003],
+    first_publish_year: 1999,
+    oclc: ['31261845'],
+  },
+  {
+    title_suggest: 'La Fin des temps',
+    edition_key: ['OL9445594M', 'OL8835168M'],
+    cover_i: 964514,
+    isbn: ['9782020134071', '2020134071', '9782020511131', '2020511134'],
+    has_fulltext: false,
+    text: [
+      'OL9445594M',
+      'OL8835168M',
+      '9782020134071',
+      '2020134071',
+      '9782020511131',
+      '2020511134',
+      'Murakami Haruki',
+      '27298504',
+      'OL382524A',
+      'La Fin des temps',
+      '/works/OL15029005W',
+      'Seuil',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL9445594M',
+      '/books/OL8835168M',
+      '/works/OL15029005W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'La Fin des temps',
+    publish_date: ['June 7, 2001', 'March 19, 1998'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 2,
+    key: '/works/OL15029005W',
+    id_goodreads: ['2899079', '122419'],
+    publisher: ['Seuil'],
+    language: ['fre'],
+    last_modified_i: 1304102238,
+    id_librarything: ['5798'],
+    cover_edition_key: 'OL8835168M',
+    publish_year: [1998, 2001],
+    first_publish_year: 1998,
+    oclc: ['27298504'],
+  },
+  {
+    title_suggest: 'Norwegian Wood',
+    edition_key: ['OL24196635M', 'OL24240971M'],
+    isbn: ['0375704027', '9780099448822', '0099448823', '9780375704024'],
+    has_fulltext: false,
+    text: [
+      'OL24196635M',
+      'OL24240971M',
+      '0375704027',
+      '9780099448822',
+      '0099448823',
+      '9780375704024',
+      'Murakami Haruki',
+      'OL382524A',
+      'Norwegian Wood',
+      '/works/OL15095603W',
+      'Vintage',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL24196635M',
+      '/books/OL24240971M',
+      '/works/OL15095603W',
+      '/authors/OL382524A',
+    ],
+    author_key: ['OL382524A'],
+    title: 'Norwegian Wood',
+    publish_date: ['2003', '2000'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['United States', 'London, United Kingdom'],
+    edition_count: 2,
+    key: '/works/OL15095603W',
+    publisher: ['Vintage'],
+    language: ['eng'],
+    last_modified_i: 1327861207,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL24196635M',
+    publish_year: [2003, 2000],
+    first_publish_year: 2000,
+  },
+  {
+    title_suggest: 'Yu tian yan tian',
+    edition_key: ['OL23061020M', 'OL23096574M'],
+    subtitle: "Xila, Tu'erqi bian jing ji xing",
+    has_fulltext: false,
+    text: [
+      'OL23061020M',
+      'OL23096574M',
+      "Xila, Tu'erqi bian jing ji xing",
+      'Murakami Haruki',
+      'Zhang, Zhibin.',
+      'Lin, Shaohua.',
+      '7532743934',
+      '9571332313',
+      '9787532743933',
+      '9789571332314',
+      'OL382524A',
+      'Yu tian yan tian',
+      'Cunshang Chunshu zhu ; Lin Shaohua yi.',
+      'Cunshang Chunshu zhu ; Zhang Zhibin yi.',
+      'Travel',
+      'Description and travel',
+      '/works/OL15338504W',
+      'Shanghai yi wen chu ban she',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+      "Xila, Tu'erqi bian jing ji xing",
+      'Uten enten.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'Haruki Murakami (1949-)',
+      'Turkey',
+      'Greece',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL23061020M',
+      '/books/OL23096574M',
+      '/works/OL15338504W',
+      '/subjects/travel',
+      '/subjects/description_and_travel',
+      '/subjects/person:haruki_murakami_(1949-)',
+      '/subjects/place:turkey',
+      '/subjects/place:greece',
+      '/authors/OL382524A',
+    ],
+    contributor: ['Zhang, Zhibin.', 'Lin, Shaohua.'],
+    isbn: ['7532743934', '9571332313', '9787532743933', '9789571332314'],
+    author_key: ['OL382524A'],
+    subject: ['Travel', 'Description and travel'],
+    title: 'Yu tian yan tian',
+    publish_date: ['2000', '2007'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['Shanghai', 'Taibei Shi'],
+    edition_count: 2,
+    key: '/works/OL15338504W',
+    publisher: [
+      'Shanghai yi wen chu ban she',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+    ],
+    language: ['chi'],
+    last_modified_i: 1282110101,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    person: ['Haruki Murakami (1949-)'],
+    publish_year: [2000, 2007],
+    first_publish_year: 2000,
+    place: ['Turkey', 'Greece'],
+  },
+  {
+    title_suggest: 'Pinball, 1973',
+    edition_key: ['OL25617142M', 'OL15303522M'],
+    isbn: ['4061860127', '9784061860124'],
+    has_fulltext: false,
+    text: [
+      'OL25617142M',
+      'OL15303522M',
+      '4061860127',
+      '9784061860124',
+      'Murakami Haruki',
+      'OL382524A',
+      'Pinball, 1973',
+      'Haruki Murakami ; translated by Alfred Birnbaum.',
+      '/works/OL2625461W',
+      'Kodansha International',
+      'Aldoguardi',
+      '1973-nen no pinboru.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL25617142M',
+      '/books/OL15303522M',
+      '/works/OL2625461W',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['2837022'],
+    author_key: ['OL382524A'],
+    title: 'Pinball, 1973',
+    publish_date: ['2014-07', '1985'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['Tokyo', 'France'],
+    edition_count: 2,
+    key: '/works/OL2625461W',
+    id_goodreads: ['591978'],
+    publisher: ['Kodansha International', 'Aldoguardi'],
+    language: ['fre', 'eng'],
+    last_modified_i: 1409042477,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1985],
+    first_publish_year: 1985,
+  },
+  {
+    title_suggest: 'KAFKA ON THE SHORE',
+    edition_key: ['OL7318410M', 'OL9273065M'],
+    cover_i: 75336,
+    isbn: ['9780099458326', '0099458322'],
+    has_fulltext: false,
+    text: [
+      'OL7318410M',
+      'OL9273065M',
+      '9780099458326',
+      '0099458322',
+      'Murakami Haruki',
+      'OL382524A',
+      'KAFKA ON THE SHORE',
+      '/works/OL7922948W',
+      'Alfred A Knopf',
+      "\"So you're all set for money, then?\" the boy named Crow asks in his characteristic sluggish voice. The kind of voice you have when you've just woken up and your mouth still feels heavy and dull. But he's just pretending. He's totally awake. As always.",
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL7318410M',
+      '/books/OL9273065M',
+      '/works/OL7922948W',
+      '/authors/OL382524A',
+    ],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'KAFKA ON THE SHORE',
+    publish_date: ['2005'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 2,
+    key: '/works/OL7922948W',
+    id_goodreads: ['818109'],
+    publisher: ['Alfred A Knopf'],
+    last_modified_i: 1325962545,
+    id_librarything: ['5819'],
+    cover_edition_key: 'OL7318410M',
+    first_sentence: [
+      "\"So you're all set for money, then?\" the boy named Crow asks in his characteristic sluggish voice. The kind of voice you have when you've just woken up and your mouth still feels heavy and dull. But he's just pretending. He's totally awake. As always.",
+    ],
+    publish_year: [2005],
+    first_publish_year: 2005,
+  },
+  {
+    title_suggest: 'Hard-Boiled Wonderland and the End of the World',
+    edition_key: ['OL9037157M', 'OL25118521M'],
+    isbn: ['9780679743460', '0-679-74346-4', '9784770018939', '4770018932'],
+    has_fulltext: true,
+    text: [
+      'OL9037157M',
+      'OL25118521M',
+      '9780679743460',
+      '0-679-74346-4',
+      '9784770018939',
+      '4770018932',
+      'Murakami Haruki',
+      'hardboiledwonder00mura_117',
+      'OL382524A',
+      'Fantasy fiction',
+      'Accessible book',
+      'Protected DAISY',
+      'Hard-Boiled Wonderland and the End of the World',
+      '/works/OL4317689W',
+      'Kodansha International',
+      'Vintage International',
+      'PL856.U673S4513 1993',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      'THE elevator continued its impossible slow ascent.',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL9037157M',
+      '/books/OL25118521M',
+      '/works/OL4317689W',
+      '/subjects/fantasy_fiction',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['5798'],
+    ia: ['hardboiledwonder00mura_117'],
+    author_key: ['OL382524A'],
+    subject: ['Fantasy fiction', 'Accessible book', 'Protected DAISY'],
+    title: 'Hard-Boiled Wonderland and the End of the World',
+    ia_collection_s: 'printdisabled;librarygenesis',
+    printdisabled_s: 'OL25118521M',
+    type: 'work',
+    ebook_count_i: 1,
+    publish_place: ['New York, USA'],
+    edition_count: 2,
+    key: '/works/OL4317689W',
+    id_goodreads: ['833604'],
+    public_scan_b: false,
+    publisher: ['Kodansha International', 'Vintage International'],
+    language: ['eng'],
+    lccn: ['PL856.U673S4513 1993'],
+    last_modified_i: 1406593983,
+    first_publish_year: 1993,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL25118521M',
+    first_sentence: ['THE elevator continued its impossible slow ascent.'],
+    publish_year: [1993],
+    publish_date: ['1993'],
+  },
+  {
+    title_suggest: 'Sputnik, Mi Amor',
+    edition_key: ['OL9144154M', 'OL25439874M'],
+    cover_i: 1048287,
+    isbn: ['8483835169', '9788483102169', '8483102161', '978-84-8383-516-6'],
+    has_fulltext: false,
+    text: [
+      'OL9144154M',
+      'OL25439874M',
+      '8483835169',
+      '9788483102169',
+      '8483102161',
+      '978-84-8383-516-6',
+      'Murakami Haruki',
+      'OL382524A',
+      'Sputnik, Mi Amor',
+      '/works/OL2625473W',
+      'TusQuets',
+      'Tusquets',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL9144154M',
+      '/books/OL25439874M',
+      '/works/OL2625473W',
+      '/authors/OL382524A',
+    ],
+    id_librarything: ['74326'],
+    author_key: ['OL382524A'],
+    title: 'Sputnik, Mi Amor',
+    publish_date: ['2009', 'November 1, 2002'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['Barcelona, Spain'],
+    edition_count: 2,
+    key: '/works/OL2625473W',
+    id_goodreads: ['17817'],
+    publisher: ['TusQuets', 'Tusquets'],
+    language: ['spa'],
+    last_modified_i: 1395761431,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    cover_edition_key: 'OL9144154M',
+    publish_year: [2002, 2009],
+    first_publish_year: 2002,
+  },
+  {
+    title_suggest: 'Kafka En La Orilla',
+    edition_key: ['OL9199671M'],
+    isbn: ['9871210558', '9789871210558'],
+    has_fulltext: false,
+    text: [
+      'OL9199671M',
+      '9871210558',
+      '9789871210558',
+      'Murakami Haruki',
+      'OL382524A',
+      'Kafka En La Orilla',
+      '/works/OL2625431W',
+      'TusQuets',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL9199671M', '/works/OL2625431W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Kafka En La Orilla',
+    publish_date: ['November 2006'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625431W',
+    id_goodreads: ['68405'],
+    publisher: ['TusQuets'],
+    language: ['spa'],
+    last_modified_i: 1281576233,
+    id_librarything: ['5819'],
+    publish_year: [2006],
+    first_publish_year: 2006,
+  },
+  {
+    title_suggest: "Lan'gehansi dao de wu hou",
+    edition_key: ['OL23017545M'],
+    isbn: ['9789571338194', '9571338192'],
+    has_fulltext: false,
+    text: [
+      'OL23017545M',
+      '9789571338194',
+      '9571338192',
+      'Murakami Haruki',
+      '52386924',
+      'OL382524A',
+      "Lan'gehansi dao de wu hou",
+      'Cunshang Chunshu zhu ; Anxi Shuiwan tu ; Zhang Zhibin yi.',
+      'Japanese essays',
+      'Translations into Chinese',
+      'Japanese literature',
+      '/works/OL2625442W',
+      'Shi bao wen hua chu ban qi ye gu fen you xian gong si',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+      '20th century',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL23017545M',
+      '/works/OL2625442W',
+      '/subjects/translations_into_chinese',
+      '/subjects/japanese_essays',
+      '/subjects/japanese_literature',
+      '/subjects/time:20th_century',
+      '/authors/OL382524A',
+    ],
+    oclc: ['52386924'],
+    author_key: ['OL382524A'],
+    subject: [
+      'Japanese essays',
+      'Translations into Chinese',
+      'Japanese literature',
+    ],
+    title: "Lan'gehansi dao de wu hou",
+    publish_date: ['2002'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['Taibei Shi'],
+    edition_count: 1,
+    key: '/works/OL2625442W',
+    publisher: ['Shi bao wen hua chu ban qi ye gu fen you xian gong si'],
+    language: ['chi'],
+    last_modified_i: 1304098112,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [2002],
+    first_publish_year: 2002,
+    time: ['20th century'],
+  },
+  {
+    title_suggest: 'Dai shu tong xin',
+    edition_key: ['OL21007386M'],
+    isbn: ['9789621714213', '9621714214'],
+    has_fulltext: false,
+    text: [
+      'OL21007386M',
+      '9789621714213',
+      '9621714214',
+      'Murakami Haruki',
+      'OL382524A',
+      'Dai shu tong xin',
+      '/works/OL2625413W',
+      'Bo Yi',
+      'Dai shu tong xin.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL21007386M', '/works/OL2625413W', '/authors/OL382524A'],
+    id_librarything: ['7758872'],
+    author_key: ['OL382524A'],
+    title: 'Dai shu tong xin',
+    publish_date: ['1995'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['Hong Kong'],
+    edition_count: 1,
+    key: '/works/OL2625413W',
+    publisher: ['Bo Yi'],
+    language: ['chi'],
+    last_modified_i: 1282157447,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1995],
+    first_publish_year: 1995,
+  },
+  {
+    title_suggest: 'Murakami karuta',
+    edition_key: ['OL23091574M'],
+    subtitle: 'usagi oish\u012b Furansujin',
+    has_fulltext: false,
+    text: [
+      'OL23091574M',
+      'usagi oish\u012b Furansujin',
+      'Murakami Haruki',
+      'Anzai, Mizumaru, 1942-',
+      '9784163689401',
+      '4163689400',
+      'OL382524A',
+      'Murakami karuta',
+      'Murakami Haruki ; Anzai Mizumaru.',
+      'Iroha karuta',
+      '/works/OL2625455W',
+      'Bungei Shunj\u016b',
+      'Usagi oish\u012b Furansujin',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL23091574M',
+      '/works/OL2625455W',
+      '/subjects/iroha_karuta',
+      '/authors/OL382524A',
+    ],
+    contributor: ['Anzai, Mizumaru, 1942-'],
+    isbn: ['9784163689401', '4163689400'],
+    author_key: ['OL382524A'],
+    subject: ['Iroha karuta'],
+    title: 'Murakami karuta',
+    publish_date: ['2007'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL2625455W',
+    id_goodreads: ['1633694'],
+    publisher: ['Bungei Shunj\u016b'],
+    language: ['jpn'],
+    last_modified_i: 1282109968,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [2007],
+    first_publish_year: 2007,
+  },
+  {
+    title_suggest: 'Wind-Up Bird Chronicles',
+    edition_key: ['OL7686139M'],
+    isbn: ['9780517414712', '0517414716'],
+    has_fulltext: false,
+    text: [
+      'OL7686139M',
+      '9780517414712',
+      '0517414716',
+      'Murakami Haruki',
+      'OL382524A',
+      'Wind-Up Bird Chronicles',
+      '/works/OL2625495W',
+      'Random House Value Publishing',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL7686139M', '/works/OL2625495W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Wind-Up Bird Chronicles',
+    publish_date: ['June 7, 1999'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625495W',
+    id_goodreads: ['11276'],
+    publisher: ['Random House Value Publishing'],
+    language: ['eng'],
+    last_modified_i: 1281070261,
+    id_librarything: ['5069'],
+    publish_year: [1999],
+    first_publish_year: 1999,
+  },
+  {
+    title_suggest: 'Kangaru-biyori',
+    edition_key: ['OL14789142M'],
+    isbn: ['9784061838581', '406183858X'],
+    has_fulltext: false,
+    text: [
+      'OL14789142M',
+      '9784061838581',
+      '406183858X',
+      'Murakami Haruki',
+      'OL382524A',
+      'Kangaru-biyori',
+      'Murakami Haruki.',
+      '/works/OL2625437W',
+      'K\u014ddansha',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL14789142M', '/works/OL2625437W', '/authors/OL382524A'],
+    id_librarything: ['8280878'],
+    author_key: ['OL382524A'],
+    title: 'Kangaru-biyori',
+    publish_date: ['1986'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL2625437W',
+    id_goodreads: ['1854919'],
+    publisher: ['K\u014ddansha'],
+    language: ['jpn'],
+    last_modified_i: 1282095101,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1986],
+    first_publish_year: 1986,
+  },
+  {
+    title_suggest: 'Ren zao wei xing qing ren',
+    publish_place: ['Xianggang'],
+    isbn: [
+      '9789621729033',
+      '9621729033',
+      '9789621789044',
+      '9621729041',
+      '9789621729040',
+      '9621789044',
+    ],
+    has_fulltext: false,
+    language: ['chi'],
+    edition_count: 1,
+    edition_key: ['OL22977194M'],
+    last_modified_i: 1282110001,
+    title: 'Ren zao wei xing qing ren',
+    publisher: ['Bo yi chu ban ji tuan you xian gong si'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1999],
+    seed: ['/books/OL22977194M', '/works/OL2625465W', '/authors/OL382524A'],
+    first_publish_year: 1999,
+    author_name: ['Murakami Haruki'],
+    key: '/works/OL2625465W',
+    text: [
+      'OL22977194M',
+      '9789621729033',
+      '9621729033',
+      '9789621789044',
+      '9621729041',
+      '9789621729040',
+      '9621789044',
+      'Murakami Haruki',
+      'OL382524A',
+      'Ren zao wei xing qing ren',
+      'Cunshang Chunshu zhu ; Lai Mingzhu yi.',
+      '/works/OL2625465W',
+      'Bo yi chu ban ji tuan you xian gong si',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    publish_date: ['1999'],
+    author_key: ['OL382524A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+  {
+    title_suggest:
+      'Wie ich eines sch\u00f6nen Morgens im April das 100%ige M\u00e4dchen sah',
+    edition_key: ['OL9048479M'],
+    cover_i: 1014239,
+    isbn: ['3499222507', '9783499222504'],
+    has_fulltext: false,
+    text: [
+      'OL9048479M',
+      '3499222507',
+      '9783499222504',
+      'Murakami Haruki',
+      'OL382524A',
+      'Wie ich eines sch\u00f6nen Morgens im April das 100%ige M\u00e4dchen sah',
+      '/works/OL2625491W',
+      'Rowohlt Tb.',
+      'Wie ich eines sch\u00f6nen Morgens im April das 100%ige M\u00e4dchen sah.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL9048479M', '/works/OL2625491W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title:
+      'Wie ich eines sch\u00f6nen Morgens im April das 100%ige M\u00e4dchen sah',
+    publish_date: ['August 1, 1998'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625491W',
+    id_goodreads: ['10134'],
+    publisher: ['Rowohlt Tb.'],
+    language: ['ger'],
+    last_modified_i: 1281453380,
+    id_librarything: ['1046456'],
+    cover_edition_key: 'OL9048479M',
+    publish_year: [1998],
+    first_publish_year: 1998,
+  },
+  {
+    title_suggest: "Wakai dokusha no tame no tanpen shosetsu an'nai",
+    edition_key: ['OL22965140M'],
+    isbn: ['4163533206', '9784163533209'],
+    has_fulltext: false,
+    text: [
+      'OL22965140M',
+      '4163533206',
+      '9784163533209',
+      'Murakami Haruki',
+      '38594148',
+      'OL382524A',
+      "Wakai dokusha no tame no tanpen shosetsu an'nai",
+      'Murakami Haruki.',
+      'Japanese fiction',
+      'History and criticism',
+      '/works/OL2625489W',
+      'Bungei Shunj\u016b',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL22965140M',
+      '/works/OL2625489W',
+      '/subjects/japanese_fiction',
+      '/subjects/history_and_criticism',
+      '/authors/OL382524A',
+    ],
+    oclc: ['38594148'],
+    author_key: ['OL382524A'],
+    subject: ['Japanese fiction', 'History and criticism'],
+    title: "Wakai dokusha no tame no tanpen shosetsu an'nai",
+    publish_date: ['1997'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL2625489W',
+    id_goodreads: ['1937497'],
+    publisher: ['Bungei Shunj\u016b'],
+    language: ['jpn'],
+    last_modified_i: 1282110054,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1997],
+    first_publish_year: 1997,
+  },
+  {
+    title_suggest: 'Murakami Asahido',
+    edition_key: ['OL12534171M'],
+    isbn: ['4948718068', '9784948718067'],
+    has_fulltext: false,
+    text: [
+      'OL12534171M',
+      '4948718068',
+      '9784948718067',
+      'Murakami Haruki',
+      '15426529',
+      'OL382524A',
+      'Murakami Asahido',
+      '/works/OL2625450W',
+      'Wakabayashi Shuppan Kikaku',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL12534171M', '/works/OL2625450W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Murakami Asahido',
+    publish_date: ['1984'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625450W',
+    id_goodreads: ['1049220'],
+    publisher: ['Wakabayashi Shuppan Kikaku'],
+    language: ['jpn'],
+    last_modified_i: 1303884661,
+    id_librarything: ['4342528'],
+    publish_year: [1984],
+    first_publish_year: 1984,
+    oclc: ['15426529'],
+  },
+  {
+    title_suggest: 'Murakami Haruki zensakuhin, 1990-2000',
+    edition_key: ['OL16664425M'],
+    isbn: ['9784061879416', '4061879413'],
+    has_fulltext: false,
+    text: [
+      'OL16664425M',
+      '9784061879416',
+      '4061879413',
+      'Murakami Haruki',
+      'OL382524A',
+      'Murakami Haruki zensakuhin, 1990-2000',
+      '[chosha Murakami Haruki].',
+      '/works/OL2625454W',
+      'K\u014ddansha',
+      '2008551626',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL16664425M', '/works/OL2625454W', '/authors/OL382524A'],
+    author_key: ['OL382524A'],
+    title: 'Murakami Haruki zensakuhin, 1990-2000',
+    publish_date: ['2002'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL2625454W',
+    id_goodreads: ['1936717'],
+    publisher: ['K\u014ddansha'],
+    language: ['jpn'],
+    lccn: ['2008551626'],
+    last_modified_i: 1282109968,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [2002],
+    first_publish_year: 2002,
+  },
+  {
+    title_suggest: 'Yoru no kumozaru',
+    edition_key: ['OL21799365M'],
+    isbn: ['9784582828894', '4582828892'],
+    has_fulltext: false,
+    text: [
+      'OL21799365M',
+      '9784582828894',
+      '4582828892',
+      'Murakami Haruki',
+      '33152848',
+      'OL382524A',
+      'Yoru no kumozaru',
+      '/works/OL2625498W',
+      'Heibon',
+      'Yoru no kumozaru.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL21799365M', '/works/OL2625498W', '/authors/OL382524A'],
+    oclc: ['33152848'],
+    author_key: ['OL382524A'],
+    title: 'Yoru no kumozaru',
+    publish_date: ['1995'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['Tokyo'],
+    edition_count: 1,
+    key: '/works/OL2625498W',
+    id_goodreads: ['1936724'],
+    publisher: ['Heibon'],
+    language: ['jpn'],
+    last_modified_i: 1303949352,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1995],
+    first_publish_year: 1995,
+  },
+  {
+    title_suggest: 'Imi ga nakereba suingu wa nai',
+    edition_key: ['OL23067977M'],
+    isbn: ['4163676007', '9784163676005'],
+    has_fulltext: false,
+    text: [
+      'OL23067977M',
+      '4163676007',
+      '9784163676005',
+      'Murakami Haruki',
+      '63790074',
+      'OL382524A',
+      'Imi ga nakereba suingu wa nai',
+      'Murakami Haruki.',
+      'Music appreciation',
+      'History and criticism',
+      'Music',
+      '/works/OL2625430W',
+      'Bungei Shunj\u016b',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: [
+      '/books/OL23067977M',
+      '/works/OL2625430W',
+      '/subjects/music_appreciation',
+      '/subjects/history_and_criticism',
+      '/subjects/music',
+      '/authors/OL382524A',
+    ],
+    oclc: ['63790074'],
+    author_key: ['OL382524A'],
+    subject: ['Music appreciation', 'History and criticism', 'Music'],
+    title: 'Imi ga nakereba suingu wa nai',
+    publish_date: ['2005'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL2625430W',
+    id_goodreads: ['1854918'],
+    publisher: ['Bungei Shunj\u016b'],
+    language: ['jpn'],
+    last_modified_i: 1304145589,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [2005],
+    first_publish_year: 2005,
+  },
+  {
+    title_suggest: 'Al Sur de La Frontera, Al Oeste del Sol',
+    edition_key: ['OL9199664M'],
+    isbn: ['9789871210091', '9871210094'],
+    has_fulltext: false,
+    text: [
+      'OL9199664M',
+      '9789871210091',
+      '9871210094',
+      'Murakami Haruki',
+      'Lourdes Porta (Translator)',
+      'OL382524A',
+      'Al Sur de La Frontera, Al Oeste del Sol',
+      '/works/OL2625405W',
+      'TusQuets',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL9199664M', '/works/OL2625405W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: ['Lourdes Porta (Translator)'],
+    author_key: ['OL382524A'],
+    title: 'Al Sur de La Frontera, Al Oeste del Sol',
+    publish_date: ['June 2005'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625405W',
+    id_goodreads: ['146930'],
+    publisher: ['TusQuets'],
+    language: ['spa'],
+    last_modified_i: 1294542721,
+    id_librarything: ['5042'],
+    publish_year: [2005],
+    first_publish_year: 2005,
+  },
+  {
+    title_suggest: 'Cronica del Pajaro Que Da Cuerda Al Mundo',
+    edition_key: ['OL9199665M'],
+    isbn: ['9871210183', '9789871210183'],
+    has_fulltext: false,
+    text: [
+      'OL9199665M',
+      '9871210183',
+      '9789871210183',
+      'Murakami Haruki',
+      '70105803',
+      'OL382524A',
+      'Cronica del Pajaro Que Da Cuerda Al Mundo',
+      '/works/OL2625412W',
+      'TusQuets',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL9199665M', '/works/OL2625412W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Cronica del Pajaro Que Da Cuerda Al Mundo',
+    publish_date: ['October 2005'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625412W',
+    id_goodreads: ['288663'],
+    publisher: ['TusQuets'],
+    language: ['spa'],
+    last_modified_i: 1304151984,
+    id_librarything: ['5069'],
+    publish_year: [2005],
+    first_publish_year: 2005,
+    oclc: ['70105803'],
+  },
+  {
+    title_suggest: 'Les Amants du Spoutnik',
+    edition_key: ['OL8932770M'],
+    cover_i: 984341,
+    isbn: ['2714438601', '9782714438607'],
+    has_fulltext: false,
+    text: [
+      'OL8932770M',
+      '2714438601',
+      '9782714438607',
+      'Murakami Haruki',
+      'OL382524A',
+      'Les Amants du Spoutnik',
+      '/works/OL2625443W',
+      'Belfond',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL8932770M', '/works/OL2625443W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Les Amants du Spoutnik',
+    publish_date: ['February 6, 2003'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625443W',
+    id_goodreads: ['122445'],
+    publisher: ['Belfond'],
+    language: ['fre'],
+    last_modified_i: 1281444776,
+    id_librarything: ['74326'],
+    cover_edition_key: 'OL8932770M',
+    publish_year: [2003],
+    first_publish_year: 2003,
+  },
+  {
+    title_suggest: 'Kafka Na Plyazhe',
+    edition_key: ['OL9045571M'],
+    isbn: ['5699109382', '9785699109388'],
+    has_fulltext: false,
+    text: [
+      'OL9045571M',
+      '5699109382',
+      '9785699109388',
+      'Murakami Haruki',
+      '57559842',
+      'OL382524A',
+      'Kafka Na Plyazhe',
+      '/works/OL2625433W',
+      'Eksmo',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL9045571M', '/works/OL2625433W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Kafka Na Plyazhe',
+    publish_date: ['2005'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625433W',
+    id_goodreads: ['463050'],
+    publisher: ['Eksmo'],
+    language: ['rus'],
+    last_modified_i: 1303805451,
+    id_librarything: ['5819'],
+    publish_year: [2005],
+    first_publish_year: 2005,
+    oclc: ['57559842'],
+  },
+  {
+    title_suggest: "Les chroniques de l'oiseau \u00e0 ressort",
+    edition_key: ['OL8834853M'],
+    cover_i: 964175,
+    isbn: ['9782020349147', '2020349140'],
+    has_fulltext: false,
+    text: [
+      'OL8834853M',
+      '9782020349147',
+      '2020349140',
+      'Murakami Haruki',
+      'OL382524A',
+      "Les chroniques de l'oiseau \u00e0 ressort",
+      '/works/OL2625444W',
+      'Seuil',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL8834853M', '/works/OL2625444W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: "Les chroniques de l'oiseau \u00e0 ressort",
+    publish_date: ['May 4, 2001'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625444W',
+    id_goodreads: ['121980'],
+    publisher: ['Seuil'],
+    language: ['fre'],
+    last_modified_i: 1281437310,
+    id_librarything: ['5069'],
+    cover_edition_key: 'OL8834853M',
+    publish_year: [2001],
+    first_publish_year: 2001,
+  },
+  {
+    title_suggest: 'Underground / Podzemka',
+    edition_key: ['OL12584784M'],
+    isbn: ['5699157700', '9785699157709'],
+    has_fulltext: false,
+    text: [
+      'OL12584784M',
+      '5699157700',
+      '9785699157709',
+      'Murakami Haruki',
+      'OL382524A',
+      'Underground / Podzemka',
+      '/works/OL2625484W',
+      'Eksmo',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL12584784M', '/works/OL2625484W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    author_key: ['OL382524A'],
+    title: 'Underground / Podzemka',
+    publish_date: ['2006'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625484W',
+    id_goodreads: ['146928'],
+    publisher: ['Eksmo'],
+    language: ['rus'],
+    last_modified_i: 1282053409,
+    id_librarything: ['58240'],
+    publish_year: [2006],
+    first_publish_year: 2006,
+  },
+  {
+    title_suggest: 'Hajimete no bungaku',
+    edition_key: ['OL23104887M'],
+    isbn: ['4163598103', '9784163598109'],
+    has_fulltext: false,
+    text: [
+      'OL23104887M',
+      '4163598103',
+      '9784163598109',
+      'Murakami Haruki',
+      '77521214',
+      'OL382524A',
+      'Hajimete no bungaku',
+      '/works/OL2625420W',
+      'Bungei Shunj\u016b',
+      'Hajimete no bungaku.',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL23104887M', '/works/OL2625420W', '/authors/OL382524A'],
+    oclc: ['77521214'],
+    author_key: ['OL382524A'],
+    title: 'Hajimete no bungaku',
+    publish_date: ['2006'],
+    type: 'work',
+    ebook_count_i: 0,
+    publish_place: ['T\u014dky\u014d'],
+    edition_count: 1,
+    key: '/works/OL2625420W',
+    id_goodreads: ['1633700'],
+    publisher: ['Bungei Shunj\u016b'],
+    language: ['jpn'],
+    last_modified_i: 1282109906,
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [2006],
+    first_publish_year: 2006,
+  },
+  {
+    title_suggest:
+      'Al Sur De La Frontera, Al Oeste Del Sol / South of the Border, West of the Sun',
+    edition_key: ['OL9144164M'],
+    cover_i: 2272110,
+    isbn: ['8483102501', '9788483102503'],
+    has_fulltext: false,
+    text: [
+      'OL9144164M',
+      '8483102501',
+      '9788483102503',
+      'Murakami Haruki',
+      'Lourdes Porta (Translator)',
+      'OL382524A',
+      'Al Sur De La Frontera, Al Oeste Del Sol / South of the Border, West of the Sun',
+      '/works/OL2625406W',
+      'TusQuets',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    author_name: ['Murakami Haruki'],
+    seed: ['/books/OL9144164M', '/works/OL2625406W', '/authors/OL382524A'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    contributor: ['Lourdes Porta (Translator)'],
+    author_key: ['OL382524A'],
+    title:
+      'Al Sur De La Frontera, Al Oeste Del Sol / South of the Border, West of the Sun',
+    publish_date: ['October 1, 2003'],
+    type: 'work',
+    ebook_count_i: 0,
+    edition_count: 1,
+    key: '/works/OL2625406W',
+    id_goodreads: ['45299'],
+    publisher: ['TusQuets'],
+    language: ['spa'],
+    last_modified_i: 1294542573,
+    id_librarything: ['5042'],
+    cover_edition_key: 'OL9144164M',
+    publish_year: [2003],
+    first_publish_year: 2003,
+  },
+  {
+    title_suggest: 'Ch\u016bgoku iki no sur\u014d b\u014dto',
+    publish_place: ['T\u014dky\u014d'],
+    has_fulltext: false,
+    language: ['jpn'],
+    edition_count: 1,
+    edition_key: ['OL22920462M'],
+    last_modified_i: 1282109886,
+    title: 'Ch\u016bgoku iki no sur\u014d b\u014dto',
+    publisher: ['Ch\u016b\u014d K\u014dronsha'],
+    author_alternative_name: ['Haruki Murakami', '\u6751\u4e0a\u6625\u6a39'],
+    publish_year: [1983],
+    seed: ['/books/OL22920462M', '/works/OL2625503W', '/authors/OL382524A'],
+    first_publish_year: 1983,
+    author_name: ['Murakami Haruki'],
+    key: '/works/OL2625503W',
+    text: [
+      'OL22920462M',
+      'Murakami Haruki',
+      'OL382524A',
+      'Ch\u016bgoku iki no sur\u014d b\u014dto',
+      'Murakami Haruki.',
+      '/works/OL2625503W',
+      'Ch\u016b\u014d K\u014dronsha',
+      'Haruki Murakami',
+      '\u6751\u4e0a\u6625\u6a39',
+    ],
+    publish_date: ['1983'],
+    author_key: ['OL382524A'],
+    type: 'work',
+    ebook_count_i: 0,
+  },
+];
